Internal function for training the BNN

This function performs the actual fitting of the Bayesian Neural Network. It is called by the exported bnns methods and is not intended for direct use.

Usage

bnns_train(
  train_x,
  train_y,
  L = 1,
  nodes = 16,
  act_fn = 2,
  out_act_fn = 1,
  iter = 1000,
  warmup = 200,
  thin = 1,
  chains = 2,
  cores = 2,
  seed = 123,
  ...
)

Arguments

train_x

A numeric matrix representing the input features (predictors) for training. Rows correspond to observations, and columns correspond to features.

train_y

A numeric vector representing the target values for training. Its length must match the number of rows in train_x.

L

An integer specifying the number of hidden layers in the neural network. Default is 1.

nodes

An integer or vector specifying the number of nodes in each hidden layer. If a single value is provided, it is applied to all layers. Default is 16.

act_fn

An integer or vector specifying the activation function(s) for the hidden layers. Options are:

1 for tanh

2 for sigmoid (default)

3 for ReLU

4 for softplus

out_act_fn

An integer specifying the activation function for the output layer. Options are:

1 for linear (default)

2 for sigmoid

3 for softmax

iter

An integer specifying the total number of iterations for the Stan sampler. Default is 1e3.

warmup

An integer specifying the number of warmup iterations for the Stan sampler. Default is 2e2.

thin

An integer specifying the thinning interval for Stan samples. Default is 1.

chains

An integer specifying the number of Markov chains. Default is 2.

cores

An integer specifying the number of CPU cores to use for parallel sampling. Default is 2.

seed

An integer specifying the random seed for reproducibility. Default is 123.

...

Additional arguments passed to the Stan sampler.

Value

An object of class "bnns" containing the following components:

fit

The fitted Stan model object.

call

The matched call.

data

A list containing the Stan data used in the model.

Details

The function uses the generate_stan_code function to dynamically generate Stan code based on the specified number of layers and nodes. Stan is then used to fit the Bayesian Neural Network.

See also

Examples

if (FALSE) { # \dontrun{
# Example usage:
train_x <- matrix(runif(100), nrow = 10, ncol = 10)
train_y <- rnorm(10)
model <- bnns:::bnns_train(train_x, train_y, L = 2, nodes = c(16, 8), act_fn = c(2, 3))

# Access Stan model fit
model$fit
} # }