Internal function to generate Stan Code Based on Output Activation Function

This function serves as a wrapper to generate Stan code for Bayesian neural networks tailored to different types of response variables. Based on the specified output activation function (out_act_fn), it delegates the code generation to the appropriate function for continuous, binary, or categorical response models.

Usage

generate_stan_code(num_layers, nodes, out_act_fn = 1)

Arguments

num_layers

An integer specifying the number of hidden layers in the neural network.

nodes

A vector of integers, where each element specifies the number of nodes in the corresponding hidden layer. The length of the vector must match num_layers.

out_act_fn

An integer specifying the output activation function, determining the type of response variable. Supported values are:

1: Continuous response (identity function as output layer).

2: Binary response (sigmoid function as output layer).

3: Categorical response (softmax function as output layer).

Value

A character string containing the Stan code for the specified Bayesian neural network model. The Stan model includes data, parameters, transformed parameters, and model blocks, adjusted based on the specified response type.

Details

This function dynamically calls one of the following functions based on the value of out_act_fn:

Continuous response: Calls generate_stan_code_cont.

Binary response: Calls generate_stan_code_bin.

Categorical response: Calls generate_stan_code_cat.

If an unsupported value is provided for out_act_fn, the function throws an error. The generated Stan code is adapted for the response type, including appropriate likelihood functions and transformations.

See also

Examples

# Generate Stan code for a continuous response model
stan_code <- generate_stan_code(num_layers = 2, nodes = c(10, 5), out_act_fn = 1)
cat(stan_code)
#> 
#> data {
#>   int<lower=1> n;
#>   int<lower=1> m;
#>   int<lower=1> L;
#>   int<lower=1> nodes[L];
#>   matrix[n, m] X;
#>   vector[n] y;
#>   int<lower=1> act_fn[L];
#> }
#> 
#> parameters {
#>   matrix[m, nodes[1]] w1;
#>   vector[nodes[1]] b1;
#>   matrix[nodes[1], nodes[2]] w2;
#>   vector[nodes[2]] b2;
#>   vector[nodes[L]] w_out;
#>   real b_out;
#>   real<lower=0> sigma;
#> }
#> 
#> transformed parameters {
#>   matrix[n, nodes[1]] z1;
#>   matrix[n, nodes[1]] a1;
#>   matrix[n, nodes[2]] z2;
#>   matrix[n, nodes[2]] a2;
#>   vector[n] y_hat;
#>   z1 = X * w1 + rep_matrix(b1', n);
#>   if (act_fn[1] == 1) a1 = tanh(z1);
#>   else if (act_fn[1] == 2) a1 = inv_logit(z1);
#>   else if (act_fn[1] == 3) a1 = log(1 + exp(z1));
#>   else if (act_fn[1] == 4) a1 = fmax(rep_matrix(0, n, nodes[1]), z1);
#>   else a1 = z1;
#>   z2 = a1 * w2 + rep_matrix(b2', n);
#>   if (act_fn[2] == 1) a2 = tanh(z2);
#>   else if (act_fn[2] == 2) a2 = inv_logit(z2);
#>   else if (act_fn[2] == 3) a2 = log(1 + exp(z2));
#>   else if (act_fn[2] == 4) a2 = fmax(rep_matrix(0, n, nodes[2]), z2);
#>   else a2 = z2;
#>   y_hat = a2 * w_out + b_out;
#> }
#> 
#> model {
#>   to_vector(w1) ~ PRIOR_WEIGHT;
#>   b1 ~ PRIOR_BIAS;
#>   to_vector(w2) ~ PRIOR_WEIGHT;
#>   b2 ~ PRIOR_BIAS;
#>   w_out ~ PRIOR_WEIGHT;
#>   b_out ~ PRIOR_BIAS;
#>   sigma ~ PRIOR_SIGMA;
#>   y ~ normal(y_hat, sigma);
#> }

# Generate Stan code for a binary response model
stan_code <- generate_stan_code(num_layers = 2, nodes = c(10, 5), out_act_fn = 2)
cat(stan_code)
#> 
#> data {
#>   int<lower=1> n;
#>   int<lower=1> m;
#>   int<lower=1> L;
#>   int<lower=1> nodes[L];
#>   matrix[n, m] X;
#>   array[n] int<lower=0, upper=1> y;
#>   int<lower=1> act_fn[L];
#> }
#> 
#> parameters {
#>   matrix[m, nodes[1]] w1;
#>   vector[nodes[1]] b1;
#>   matrix[nodes[1], nodes[2]] w2;
#>   vector[nodes[2]] b2;
#>   vector[nodes[L]] w_out;
#>   real b_out;
#> }
#> 
#> transformed parameters {
#>   matrix[n, nodes[1]] z1;
#>   matrix[n, nodes[1]] a1;
#>   matrix[n, nodes[2]] z2;
#>   matrix[n, nodes[2]] a2;
#>   vector[n] y_hat;
#>   z1 = X * w1 + rep_matrix(b1', n);
#>   if (act_fn[1] == 1) a1 = tanh(z1);
#>   else if (act_fn[1] == 2) a1 = inv_logit(z1);
#>   else if (act_fn[1] == 3) a1 = log(1 + exp(z1));
#>   else if (act_fn[1] == 4) a1 = fmax(rep_matrix(0, n, nodes[1]), z1);
#>   else a1 = z1;
#>   z2 = a1 * w2 + rep_matrix(b2', n);
#>   if (act_fn[2] == 1) a2 = tanh(z2);
#>   else if (act_fn[2] == 2) a2 = inv_logit(z2);
#>   else if (act_fn[2] == 3) a2 = log(1 + exp(z2));
#>   else if (act_fn[2] == 4) a2 = fmax(rep_matrix(0, n, nodes[2]), z2);
#>   else a2 = z2;
#>   y_hat = a2 * w_out + b_out;
#> }
#> 
#> model {
#>   to_vector(w1) ~ PRIOR_WEIGHT;
#>   b1 ~ PRIOR_BIAS;
#>   to_vector(w2) ~ PRIOR_WEIGHT;
#>   b2 ~ PRIOR_BIAS;
#>   w_out ~ PRIOR_WEIGHT;
#>   b_out ~ PRIOR_BIAS;
#>   y ~ bernoulli_logit(y_hat);
#> }

# Generate Stan code for a categorical response model
stan_code <- generate_stan_code(num_layers = 2, nodes = c(10, 5), out_act_fn = 3)
cat(stan_code)
#> 
#> data {
#>   int<lower=1> n;
#>   int<lower=1> m;
#>   int<lower=1> L;
#>   int<lower=1> nodes[L];
#>   matrix[n, m] X;
#>   array[n] int<lower=1> y;
#>   int<lower=1> act_fn[L];
#>   int<lower=2> K; // Number of categories
#> }
#> 
#> parameters {
#>   matrix[m, nodes[1]] w1;
#>   vector[nodes[1]] b1;
#>   matrix[nodes[1], nodes[2]] w2;
#>   vector[nodes[2]] b2;
#>   matrix[nodes[L], K] w_out;
#>   vector[K] b_out;
#> }
#> 
#> transformed parameters {
#>   matrix[n, nodes[1]] z1;
#>   matrix[n, nodes[1]] a1;
#>   matrix[n, nodes[2]] z2;
#>   matrix[n, nodes[2]] a2;
#>   matrix[n, K] y_hat;
#>   z1 = X * w1 + rep_matrix(b1', n);
#>   if (act_fn[1] == 1) a1 = tanh(z1);
#>   else if (act_fn[1] == 2) a1 = inv_logit(z1);
#>   else if (act_fn[1] == 3) a1 = log(1 + exp(z1));
#>   else if (act_fn[1] == 4) a1 = fmax(rep_matrix(0, n, nodes[1]), z1);
#>   else a1 = z1;
#>   z2 = a1 * w2 + rep_matrix(b2', n);
#>   if (act_fn[2] == 1) a2 = tanh(z2);
#>   else if (act_fn[2] == 2) a2 = inv_logit(z2);
#>   else if (act_fn[2] == 3) a2 = log(1 + exp(z2));
#>   else if (act_fn[2] == 4) a2 = fmax(rep_matrix(0, n, nodes[2]), z2);
#>   else a2 = z2;
#>   y_hat = a2 * w_out + rep_matrix(b_out', n);
#> }
#> 
#> model {
#>   to_vector(w1) ~ PRIOR_WEIGHT;
#>   b1 ~ PRIOR_BIAS;
#>   to_vector(w2) ~ PRIOR_WEIGHT;
#>   b2 ~ PRIOR_BIAS;
#>   to_vector(w_out) ~ PRIOR_WEIGHT;
#>   b_out ~ PRIOR_BIAS;
#>   for (i in 1:n) y[i] ~ categorical_logit(y_hat[i]');
#> }