Generate Stan Code for Neural Networks with Categorical Response

This function generates Stan code for modeling a categorical response using neural networks with multiple layers. The generated code supports customizable activation functions for each layer and softmax-based prediction for the categorical output.

Usage

generate_stan_code_cat(num_layers, nodes)

Arguments

num_layers

Integer. Number of layers in the neural network.

nodes

Integer vector. Number of nodes in each layer. The length of this vector must match num_layers, and all values must be positive.

Value

A string containing the Stan code for the specified neural network architecture and categorical response model.

Details

The Stan code includes the following components:

Data Block: Defines inputs, response variable, layer configurations, and activation functions.

Parameters Block: Declares weights and biases for all layers and the output layer.

Transformed Parameters Block: Computes intermediate outputs (z and a) for each layer and calculates the final predictions (y_hat) using the softmax function.

Model Block: Specifies priors for parameters and models the categorical response using categorical_logit.

Supported activation functions for the hidden layers:

1: Tanh

2: Sigmoid

3: Softplus

4: ReLU

5: linear

The categorical response (y) is assumed to take integer values from 1 to K, where K is the total number of categories.

See also

Examples

# Generate Stan code for a neural network with 3 layers
num_layers <- 3
nodes <- c(10, 8, 6) # 10 nodes in the first layer, 8 in the second, 6 in the third
stan_code <- generate_stan_code_cat(num_layers, nodes)
cat(stan_code)
#> 
#> data {
#>   int<lower=1> n;
#>   int<lower=1> m;
#>   int<lower=1> L;
#>   int<lower=1> nodes[L];
#>   matrix[n, m] X;
#>   array[n] int<lower=1> y;
#>   int<lower=1> act_fn[L];
#>   int<lower=2> K; // Number of categories
#> }
#> 
#> parameters {
#>   matrix[m, nodes[1]] w1;
#>   vector[nodes[1]] b1;
#>   matrix[nodes[1], nodes[2]] w2;
#>   vector[nodes[2]] b2;
#>   matrix[nodes[2], nodes[3]] w3;
#>   vector[nodes[3]] b3;
#>   matrix[nodes[L], K] w_out;
#>   vector[K] b_out;
#> }
#> 
#> transformed parameters {
#>   matrix[n, nodes[1]] z1;
#>   matrix[n, nodes[1]] a1;
#>   matrix[n, nodes[2]] z2;
#>   matrix[n, nodes[2]] a2;
#>   matrix[n, nodes[3]] z3;
#>   matrix[n, nodes[3]] a3;
#>   matrix[n, K] y_hat;
#>   z1 = X * w1 + rep_matrix(b1', n);
#>   if (act_fn[1] == 1) a1 = tanh(z1);
#>   else if (act_fn[1] == 2) a1 = inv_logit(z1);
#>   else if (act_fn[1] == 3) a1 = log(1 + exp(z1));
#>   else if (act_fn[1] == 4) a1 = fmax(rep_matrix(0, n, nodes[1]), z1);
#>   else a1 = z1;
#>   z2 = a1 * w2 + rep_matrix(b2', n);
#>   if (act_fn[2] == 1) a2 = tanh(z2);
#>   else if (act_fn[2] == 2) a2 = inv_logit(z2);
#>   else if (act_fn[2] == 3) a2 = log(1 + exp(z2));
#>   else if (act_fn[2] == 4) a2 = fmax(rep_matrix(0, n, nodes[2]), z2);
#>   else a2 = z2;
#>   z3 = a2 * w3 + rep_matrix(b3', n);
#>   if (act_fn[3] == 1) a3 = tanh(z3);
#>   else if (act_fn[3] == 2) a3 = inv_logit(z3);
#>   else if (act_fn[3] == 3) a3 = log(1 + exp(z3));
#>   else if (act_fn[3] == 4) a3 = fmax(rep_matrix(0, n, nodes[3]), z3);
#>   else a3 = z3;
#>   y_hat = a3 * w_out + rep_matrix(b_out', n);
#> }
#> 
#> model {
#>   to_vector(w1) ~ PRIOR_WEIGHT;
#>   b1 ~ PRIOR_BIAS;
#>   to_vector(w2) ~ PRIOR_WEIGHT;
#>   b2 ~ PRIOR_BIAS;
#>   to_vector(w3) ~ PRIOR_WEIGHT;
#>   b3 ~ PRIOR_BIAS;
#>   to_vector(w_out) ~ PRIOR_WEIGHT;
#>   b_out ~ PRIOR_BIAS;
#>   for (i in 1:n) y[i] ~ categorical_logit(y_hat[i]');
#> }