Bayesian Neural Network Model Using Formula(default) Interface

Fits a Bayesian Neural Network (BNN) model using a formula interface. The function parses the formula and data to create the input feature matrix and target vector, then fits the model using bnns.default.

Usage

# Default S3 method
bnns(
  formula,
  data = list(),
  L = 1,
  nodes = 16,
  act_fn = 2,
  out_act_fn = 1,
  iter = 1000,
  warmup = 200,
  thin = 1,
  chains = 2,
  cores = 2,
  seed = 123,
  prior_weights = NULL,
  prior_sigma = NULL,
  ...
)

Arguments

formula

A symbolic description of the model to be fitted. The formula should specify the response variable and predictors (e.g., y ~ x1 + x2).

data

A data frame or list containing the variables in the model. Default is an empty list.

L

An integer specifying the number of hidden layers in the neural network. Default is 1.

nodes

An integer or vector specifying the number of nodes in each hidden layer. If a single value is provided, it is applied to all layers. Default is 16.

act_fn

An integer or vector specifying the activation function(s) for the hidden layers. Options are:

1 for tanh

2 for sigmoid (default)

3 for ReLU

4 for softplus

out_act_fn

An integer specifying the activation function for the output layer. Options are:

1 for linear (default)

2 for sigmoid

3 for softmax

iter

An integer specifying the total number of iterations for the Stan sampler. Default is 1e3.

warmup

An integer specifying the number of warmup iterations for the Stan sampler. Default is 2e2.

thin

An integer specifying the thinning interval for Stan samples. Default is 1.

chains

An integer specifying the number of Markov chains. Default is 2.

cores

An integer specifying the number of CPU cores to use for parallel sampling. Default is 2.

seed

An integer specifying the random seed for reproducibility. Default is 123.

prior_weights

A list specifying the prior distribution for the weights in the neural network. The list must include two components:

dist: A character string specifying the distribution type. Supported values are "normal", "uniform", and "cauchy".

params: A named list specifying the parameters for the chosen distribution:

For "normal": Provide mean (mean of the distribution) and sd (standard deviation).

For "uniform": Provide alpha (lower bound) and beta (upper bound).

For "cauchy": Provide mu (location parameter) and sigma (scale parameter).

If prior_weights is NULL, the default prior is a normal(0, 1) distribution. For example:

list(dist = "normal", params = list(mean = 0, sd = 1))

list(dist = "uniform", params = list(alpha = -1, beta = 1))

list(dist = "cauchy", params = list(mu = 0, sigma = 2.5))

prior_sigma

A list specifying the prior distribution for the sigma parameter in regression models (out_act_fn = 1). This allows for setting priors on the standard deviation of the residuals. The list must include two components:

dist: A character string specifying the distribution type. Supported values are "half-normal" and "inverse-gamma".

params: A named list specifying the parameters for the chosen distribution:

For "half-normal": Provide sd (standard deviation of the half-normal distribution).

For "inverse-gamma": Provide shape (shape parameter) and scale (scale parameter).

If prior_sigma is NULL, the default prior is a half-normal(0, 1) distribution. For example:

list(dist = "half_normal", params = list(mean = 0, sd = 1))

list(dist = "inv_gamma", params = list(alpha = 1, beta = 1))

...

Currently not in use.

Value

An object of class "bnns" containing the fitted model and associated information, including:

fit: The fitted Stan model object.

data: A list containing the processed training data.

call: The matched function call.

formula: The formula used for the model.

Details

The function uses the provided formula and data to generate the design matrix for the predictors and the response vector. It then calls helper function bnns_train to fit the Bayesian Neural Network model.

Examples

# Example usage with formula interface:
data <- data.frame(x1 = runif(10), x2 = runif(10), y = rnorm(10))
model <- bnns(y ~ -1 + x1 + x2,
  data = data, L = 1, nodes = 2, act_fn = 3,
  iter = 1e2, warmup = 5e1, chains = 1
)
#> 
#> TRANSLATING MODEL '' FROM Stan CODE TO C++ CODE NOW.
#> OS: x86_64, linux-gnu; rstan: 2.32.6; Rcpp: 1.0.13.1; inline: 0.3.20 
#>  >> setting environment variables: 
#> PKG_LIBS =  '/home/runner/work/_temp/Library/rstan/lib//libStanServices.a' -L'/home/runner/work/_temp/Library/StanHeaders/lib/' -lStanHeaders -L'/home/runner/work/_temp/Library/RcppParallel/lib/' -ltbb 
#> PKG_CPPFLAGS =   -I"/home/runner/work/_temp/Library/Rcpp/include/"  -I"/home/runner/work/_temp/Library/RcppEigen/include/"  -I"/home/runner/work/_temp/Library/RcppEigen/include/unsupported"  -I"/home/runner/work/_temp/Library/BH/include" -I"/home/runner/work/_temp/Library/StanHeaders/include/src/"  -I"/home/runner/work/_temp/Library/StanHeaders/include/"  -I"/home/runner/work/_temp/Library/RcppParallel/include/"  -I"/home/runner/work/_temp/Library/rstan/include" -DEIGEN_NO_DEBUG  -DBOOST_DISABLE_ASSERTS  -DBOOST_PENDING_INTEGER_LOG2_HPP  -DSTAN_THREADS  -DUSE_STANC3 -DSTRICT_R_HEADERS  -DBOOST_PHOENIX_NO_VARIADIC_EXPRESSION  -D_HAS_AUTO_PTR_ETC=0  -include '/home/runner/work/_temp/Library/StanHeaders/include/stan/math/prim/fun/Eigen.hpp'  -D_REENTRANT -DRCPP_PARALLEL_USE_TBB=1 
#>  >> Program source :
#> 
#>    1 : 
#>    2 : // includes from the plugin
#>    3 : // [[Rcpp::plugins(cpp14)]]
#>    4 : 
#>    5 : 
#>    6 : // user includes
#>    7 : #include <Rcpp.h>
#>    8 : using namespace Rcpp;
#>    9 : #ifndef MODELS_HPP
#>   10 : #define MODELS_HPP
#>   11 : #define STAN__SERVICES__COMMAND_HPP
#>   12 : #include <rstan/rstaninc.hpp>
#>   13 : #ifndef USE_STANC3
#>   14 : #define USE_STANC3
#>   15 : #endif
#>   16 : // Code generated by stanc v2.32.2
#>   17 : #include <stan/model/model_header.hpp>
#>   18 : namespace model1bc35ea325f1__namespace {
#>   19 : using stan::model::model_base_crtp;
#>   20 : using namespace stan::math;
#>   21 : stan::math::profile_map profiles__;
#>   22 : static constexpr std::array<const char*, 43> locations_array__ =
#>   23 :   {" (found before start of program)",
#>   24 :   " (in 'anon_model', line 11, column 2 to column 22)",
#>   25 :   " (in 'anon_model', line 12, column 2 to column 19)",
#>   26 :   " (in 'anon_model', line 13, column 2 to column 22)",
#>   27 :   " (in 'anon_model', line 14, column 2 to column 13)",
#>   28 :   " (in 'anon_model', line 15, column 2 to column 22)",
#>   29 :   " (in 'anon_model', line 18, column 2 to column 22)",
#>   30 :   " (in 'anon_model', line 19, column 2 to column 22)",
#>   31 :   " (in 'anon_model', line 20, column 2 to column 18)",
#>   32 :   " (in 'anon_model', line 21, column 2 to column 35)",
#>   33 :   " (in 'anon_model', line 25, column 7 to column 46)",
#>   34 :   " (in 'anon_model', line 24, column 24 to column 46)",
#>   35 :   " (in 'anon_model', line 24, column 7 to line 25, column 46)",
#>   36 :   " (in 'anon_model', line 23, column 24 to column 43)",
#>   37 :   " (in 'anon_model', line 23, column 7 to line 25, column 46)",
#>   38 :   " (in 'anon_model', line 22, column 19 to column 33)",
#>   39 :   " (in 'anon_model', line 22, column 2 to line 25, column 46)",
#>   40 :   " (in 'anon_model', line 26, column 2 to column 29)",
#>   41 :   " (in 'anon_model', line 29, column 2 to column 45)",
#>   42 :   " (in 'anon_model', line 30, column 2 to column 34)",
#>   43 :   " (in 'anon_model', line 31, column 2 to column 37)",
#>   44 :   " (in 'anon_model', line 32, column 2 to column 37)",
#>   45 :   " (in 'anon_model', line 33, column 2 to column 37)",
#>   46 :   " (in 'anon_model', line 34, column 2 to column 27)",
#>   47 :   " (in 'anon_model', line 2, column 2 to column 17)",
#>   48 :   " (in 'anon_model', line 3, column 2 to column 17)",
#>   49 :   " (in 'anon_model', line 4, column 2 to column 17)",
#>   50 :   " (in 'anon_model', line 5, column 2 to column 21)",
#>   51 :   " (in 'anon_model', line 6, column 9 to column 10)",
#>   52 :   " (in 'anon_model', line 6, column 12 to column 13)",
#>   53 :   " (in 'anon_model', line 6, column 2 to column 17)",
#>   54 :   " (in 'anon_model', line 7, column 9 to column 10)",
#>   55 :   " (in 'anon_model', line 7, column 2 to column 14)",
#>   56 :   " (in 'anon_model', line 8, column 2 to column 31)",
#>   57 :   " (in 'anon_model', line 11, column 9 to column 10)",
#>   58 :   " (in 'anon_model', line 11, column 12 to column 17)",
#>   59 :   " (in 'anon_model', line 12, column 9 to column 14)",
#>   60 :   " (in 'anon_model', line 13, column 9 to column 14)",
#>   61 :   " (in 'anon_model', line 18, column 9 to column 10)",
#>   62 :   " (in 'anon_model', line 18, column 12 to column 17)",
#>   63 :   " (in 'anon_model', line 19, column 9 to column 10)",
#>   64 :   " (in 'anon_model', line 19, column 12 to column 17)",
#>   65 :   " (in 'anon_model', line 20, column 9 to column 10)"};
#>   66 : class model1bc35ea325f1_ final : public model_base_crtp<model1bc35ea325f1_> {
#>   67 : private:
#>   68 :   int n;
#>   69 :   int m;
#>   70 :   int L;
#>   71 :   int nodes;
#>   72 :   Eigen::Matrix<double,-1,-1> X_data__;
#>   73 :   Eigen::Matrix<double,-1,1> y_data__;
#>   74 :   int act_fn;
#>   75 :   Eigen::Map<Eigen::Matrix<double,-1,-1>> X{nullptr, 0, 0};
#>   76 :   Eigen::Map<Eigen::Matrix<double,-1,1>> y{nullptr, 0};
#>   77 : public:
#>   78 :   ~model1bc35ea325f1_() {}
#>   79 :   model1bc35ea325f1_(stan::io::var_context& context__, unsigned int
#>   80 :                      random_seed__ = 0, std::ostream* pstream__ = nullptr)
#>   81 :       : model_base_crtp(0) {
#>   82 :     int current_statement__ = 0;
#>   83 :     using local_scalar_t__ = double;
#>   84 :     boost::ecuyer1988 base_rng__ =
#>   85 :       stan::services::util::create_rng(random_seed__, 0);
#>   86 :     // suppress unused var warning
#>   87 :     (void) base_rng__;
#>   88 :     static constexpr const char* function__ =
#>   89 :       "model1bc35ea325f1__namespace::model1bc35ea325f1_";
#>   90 :     // suppress unused var warning
#>   91 :     (void) function__;
#>   92 :     local_scalar_t__ DUMMY_VAR__(std::numeric_limits<double>::quiet_NaN());
#>   93 :     // suppress unused var warning
#>   94 :     (void) DUMMY_VAR__;
#>   95 :     try {
#>   96 :       int pos__ = std::numeric_limits<int>::min();
#>   97 :       pos__ = 1;
#>   98 :       current_statement__ = 24;
#>   99 :       context__.validate_dims("data initialization", "n", "int",
#>  100 :         std::vector<size_t>{});
#>  101 :       n = std::numeric_limits<int>::min();
#>  102 :       current_statement__ = 24;
#>  103 :       n = context__.vals_i("n")[(1 - 1)];
#>  104 :       current_statement__ = 24;
#>  105 :       stan::math::check_greater_or_equal(function__, "n", n, 1);
#>  106 :       current_statement__ = 25;
#>  107 :       context__.validate_dims("data initialization", "m", "int",
#>  108 :         std::vector<size_t>{});
#>  109 :       m = std::numeric_limits<int>::min();
#>  110 :       current_statement__ = 25;
#>  111 :       m = context__.vals_i("m")[(1 - 1)];
#>  112 :       current_statement__ = 25;
#>  113 :       stan::math::check_greater_or_equal(function__, "m", m, 1);
#>  114 :       current_statement__ = 26;
#>  115 :       context__.validate_dims("data initialization", "L", "int",
#>  116 :         std::vector<size_t>{});
#>  117 :       L = std::numeric_limits<int>::min();
#>  118 :       current_statement__ = 26;
#>  119 :       L = context__.vals_i("L")[(1 - 1)];
#>  120 :       current_statement__ = 26;
#>  121 :       stan::math::check_greater_or_equal(function__, "L", L, 1);
#>  122 :       current_statement__ = 27;
#>  123 :       context__.validate_dims("data initialization", "nodes", "int",
#>  124 :         std::vector<size_t>{});
#>  125 :       nodes = std::numeric_limits<int>::min();
#>  126 :       current_statement__ = 27;
#>  127 :       nodes = context__.vals_i("nodes")[(1 - 1)];
#>  128 :       current_statement__ = 27;
#>  129 :       stan::math::check_greater_or_equal(function__, "nodes", nodes, 1);
#>  130 :       current_statement__ = 28;
#>  131 :       stan::math::validate_non_negative_index("X", "n", n);
#>  132 :       current_statement__ = 29;
#>  133 :       stan::math::validate_non_negative_index("X", "m", m);
#>  134 :       current_statement__ = 30;
#>  135 :       context__.validate_dims("data initialization", "X", "double",
#>  136 :         std::vector<size_t>{static_cast<size_t>(n), static_cast<size_t>(m)});
#>  137 :       X_data__ = Eigen::Matrix<double,-1,-1>::Constant(n, m,
#>  138 :                    std::numeric_limits<double>::quiet_NaN());
#>  139 :       new (&X) Eigen::Map<Eigen::Matrix<double,-1,-1>>(X_data__.data(), n, m);
#>  140 :       {
#>  141 :         std::vector<local_scalar_t__> X_flat__;
#>  142 :         current_statement__ = 30;
#>  143 :         X_flat__ = context__.vals_r("X");
#>  144 :         current_statement__ = 30;
#>  145 :         pos__ = 1;
#>  146 :         current_statement__ = 30;
#>  147 :         for (int sym1__ = 1; sym1__ <= m; ++sym1__) {
#>  148 :           current_statement__ = 30;
#>  149 :           for (int sym2__ = 1; sym2__ <= n; ++sym2__) {
#>  150 :             current_statement__ = 30;
#>  151 :             stan::model::assign(X, X_flat__[(pos__ - 1)],
#>  152 :               "assigning variable X", stan::model::index_uni(sym2__),
#>  153 :               stan::model::index_uni(sym1__));
#>  154 :             current_statement__ = 30;
#>  155 :             pos__ = (pos__ + 1);
#>  156 :           }
#>  157 :         }
#>  158 :       }
#>  159 :       current_statement__ = 31;
#>  160 :       stan::math::validate_non_negative_index("y", "n", n);
#>  161 :       current_statement__ = 32;
#>  162 :       context__.validate_dims("data initialization", "y", "double",
#>  163 :         std::vector<size_t>{static_cast<size_t>(n)});
#>  164 :       y_data__ = Eigen::Matrix<double,-1,1>::Constant(n,
#>  165 :                    std::numeric_limits<double>::quiet_NaN());
#>  166 :       new (&y) Eigen::Map<Eigen::Matrix<double,-1,1>>(y_data__.data(), n);
#>  167 :       {
#>  168 :         std::vector<local_scalar_t__> y_flat__;
#>  169 :         current_statement__ = 32;
#>  170 :         y_flat__ = context__.vals_r("y");
#>  171 :         current_statement__ = 32;
#>  172 :         pos__ = 1;
#>  173 :         current_statement__ = 32;
#>  174 :         for (int sym1__ = 1; sym1__ <= n; ++sym1__) {
#>  175 :           current_statement__ = 32;
#>  176 :           stan::model::assign(y, y_flat__[(pos__ - 1)],
#>  177 :             "assigning variable y", stan::model::index_uni(sym1__));
#>  178 :           current_statement__ = 32;
#>  179 :           pos__ = (pos__ + 1);
#>  180 :         }
#>  181 :       }
#>  182 :       current_statement__ = 33;
#>  183 :       context__.validate_dims("data initialization", "act_fn", "int",
#>  184 :         std::vector<size_t>{});
#>  185 :       act_fn = std::numeric_limits<int>::min();
#>  186 :       current_statement__ = 33;
#>  187 :       act_fn = context__.vals_i("act_fn")[(1 - 1)];
#>  188 :       current_statement__ = 33;
#>  189 :       stan::math::check_greater_or_equal(function__, "act_fn", act_fn, 1);
#>  190 :       current_statement__ = 33;
#>  191 :       stan::math::check_less_or_equal(function__, "act_fn", act_fn, 3);
#>  192 :       current_statement__ = 34;
#>  193 :       stan::math::validate_non_negative_index("w1", "m", m);
#>  194 :       current_statement__ = 35;
#>  195 :       stan::math::validate_non_negative_index("w1", "nodes", nodes);
#>  196 :       current_statement__ = 36;
#>  197 :       stan::math::validate_non_negative_index("b1", "nodes", nodes);
#>  198 :       current_statement__ = 37;
#>  199 :       stan::math::validate_non_negative_index("w_out", "nodes", nodes);
#>  200 :       current_statement__ = 38;
#>  201 :       stan::math::validate_non_negative_index("z1", "n", n);
#>  202 :       current_statement__ = 39;
#>  203 :       stan::math::validate_non_negative_index("z1", "nodes", nodes);
#>  204 :       current_statement__ = 40;
#>  205 :       stan::math::validate_non_negative_index("a1", "n", n);
#>  206 :       current_statement__ = 41;
#>  207 :       stan::math::validate_non_negative_index("a1", "nodes", nodes);
#>  208 :       current_statement__ = 42;
#>  209 :       stan::math::validate_non_negative_index("y_hat", "n", n);
#>  210 :     } catch (const std::exception& e) {
#>  211 :       stan::lang::rethrow_located(e, locations_array__[current_statement__]);
#>  212 :     }
#>  213 :     num_params_r__ = (m * nodes) + nodes + nodes + 1 + 1;
#>  214 :   }
#>  215 :   inline std::string model_name() const final {
#>  216 :     return "model1bc35ea325f1_";
#>  217 :   }
#>  218 :   inline std::vector<std::string> model_compile_info() const noexcept {
#>  219 :     return std::vector<std::string>{"stanc_version = stanc3 v2.32.2",
#>  220 :              "stancflags = --"};
#>  221 :   }
#>  222 :   template <bool propto__, bool jacobian__, typename VecR, typename VecI,
#>  223 :             stan::require_vector_like_t<VecR>* = nullptr,
#>  224 :             stan::require_vector_like_vt<std::is_integral, VecI>* = nullptr>
#>  225 :   inline stan::scalar_type_t<VecR>
#>  226 :   log_prob_impl(VecR& params_r__, VecI& params_i__, std::ostream*
#>  227 :                 pstream__ = nullptr) const {
#>  228 :     using T__ = stan::scalar_type_t<VecR>;
#>  229 :     using local_scalar_t__ = T__;
#>  230 :     T__ lp__(0.0);
#>  231 :     stan::math::accumulator<T__> lp_accum__;
#>  232 :     stan::io::deserializer<local_scalar_t__> in__(params_r__, params_i__);
#>  233 :     int current_statement__ = 0;
#>  234 :     local_scalar_t__ DUMMY_VAR__(std::numeric_limits<double>::quiet_NaN());
#>  235 :     // suppress unused var warning
#>  236 :     (void) DUMMY_VAR__;
#>  237 :     static constexpr const char* function__ =
#>  238 :       "model1bc35ea325f1__namespace::log_prob";
#>  239 :     // suppress unused var warning
#>  240 :     (void) function__;
#>  241 :     try {
#>  242 :       Eigen::Matrix<local_scalar_t__,-1,-1> w1 =
#>  243 :         Eigen::Matrix<local_scalar_t__,-1,-1>::Constant(m, nodes,
#>  244 :           DUMMY_VAR__);
#>  245 :       current_statement__ = 1;
#>  246 :       w1 = in__.template read<Eigen::Matrix<local_scalar_t__,-1,-1>>(m,
#>  247 :              nodes);
#>  248 :       Eigen::Matrix<local_scalar_t__,-1,1> b1 =
#>  249 :         Eigen::Matrix<local_scalar_t__,-1,1>::Constant(nodes, DUMMY_VAR__);
#>  250 :       current_statement__ = 2;
#>  251 :       b1 = in__.template read<Eigen::Matrix<local_scalar_t__,-1,1>>(nodes);
#>  252 :       Eigen::Matrix<local_scalar_t__,-1,1> w_out =
#>  253 :         Eigen::Matrix<local_scalar_t__,-1,1>::Constant(nodes, DUMMY_VAR__);
#>  254 :       current_statement__ = 3;
#>  255 :       w_out = in__.template read<Eigen::Matrix<local_scalar_t__,-1,1>>(nodes);
#>  256 :       local_scalar_t__ b_out = DUMMY_VAR__;
#>  257 :       current_statement__ = 4;
#>  258 :       b_out = in__.template read<local_scalar_t__>();
#>  259 :       local_scalar_t__ sigma = DUMMY_VAR__;
#>  260 :       current_statement__ = 5;
#>  261 :       sigma = in__.template read_constrain_lb<local_scalar_t__,
#>  262 :                 jacobian__>(0, lp__);
#>  263 :       Eigen::Matrix<local_scalar_t__,-1,-1> z1 =
#>  264 :         Eigen::Matrix<local_scalar_t__,-1,-1>::Constant(n, nodes,
#>  265 :           DUMMY_VAR__);
#>  266 :       Eigen::Matrix<local_scalar_t__,-1,-1> a1 =
#>  267 :         Eigen::Matrix<local_scalar_t__,-1,-1>::Constant(n, nodes,
#>  268 :           DUMMY_VAR__);
#>  269 :       Eigen::Matrix<local_scalar_t__,-1,1> y_hat =
#>  270 :         Eigen::Matrix<local_scalar_t__,-1,1>::Constant(n, DUMMY_VAR__);
#>  271 :       current_statement__ = 9;
#>  272 :       stan::model::assign(z1,
#>  273 :         stan::math::add(stan::math::multiply(X, w1),
#>  274 :           stan::math::rep_matrix(stan::math::transpose(b1), n)),
#>  275 :         "assigning variable z1");
#>  276 :       current_statement__ = 16;
#>  277 :       if (stan::math::logical_eq(act_fn, 1)) {
#>  278 :         current_statement__ = 15;
#>  279 :         stan::model::assign(a1, stan::math::tanh(z1), "assigning variable a1");
#>  280 :       } else {
#>  281 :         current_statement__ = 14;
#>  282 :         if (stan::math::logical_eq(act_fn, 2)) {
#>  283 :           current_statement__ = 13;
#>  284 :           stan::model::assign(a1, stan::math::inv_logit(z1),
#>  285 :             "assigning variable a1");
#>  286 :         } else {
#>  287 :           current_statement__ = 12;
#>  288 :           if (stan::math::logical_eq(act_fn, 3)) {
#>  289 :             current_statement__ = 11;
#>  290 :             stan::model::assign(a1,
#>  291 :               stan::math::log(stan::math::add(1, stan::math::exp(z1))),
#>  292 :               "assigning variable a1");
#>  293 :           } else {
#>  294 :             current_statement__ = 10;
#>  295 :             stan::model::assign(a1,
#>  296 :               stan::math::fmax(stan::math::rep_matrix(0, n, nodes), z1),
#>  297 :               "assigning variable a1");
#>  298 :           }
#>  299 :         }
#>  300 :       }
#>  301 :       current_statement__ = 17;
#>  302 :       stan::model::assign(y_hat,
#>  303 :         stan::math::add(stan::math::multiply(a1, w_out), b_out),
#>  304 :         "assigning variable y_hat");
#>  305 :       {
#>  306 :         current_statement__ = 18;
#>  307 :         lp_accum__.add(stan::math::normal_lpdf<propto__>(
#>  308 :                          stan::math::to_vector(w1), 0.000000, 1.000000));
#>  309 :         current_statement__ = 19;
#>  310 :         lp_accum__.add(stan::math::normal_lpdf<propto__>(b1, 0.000000,
#>  311 :                          1.000000));
#>  312 :         current_statement__ = 20;
#>  313 :         lp_accum__.add(stan::math::normal_lpdf<propto__>(w_out, 0.000000,
#>  314 :                          1.000000));
#>  315 :         current_statement__ = 21;
#>  316 :         lp_accum__.add(stan::math::normal_lpdf<propto__>(b_out, 0.000000,
#>  317 :                          1.000000));
#>  318 :         current_statement__ = 22;
#>  319 :         lp_accum__.add(stan::math::normal_lpdf<propto__>(sigma, 0.000000,
#>  320 :                          1.000000));
#>  321 :         current_statement__ = 23;
#>  322 :         lp_accum__.add(stan::math::normal_lpdf<propto__>(y, y_hat, sigma));
#>  323 :       }
#>  324 :     } catch (const std::exception& e) {
#>  325 :       stan::lang::rethrow_located(e, locations_array__[current_statement__]);
#>  326 :     }
#>  327 :     lp_accum__.add(lp__);
#>  328 :     return lp_accum__.sum();
#>  329 :   }
#>  330 :   template <typename RNG, typename VecR, typename VecI, typename VecVar,
#>  331 :             stan::require_vector_like_vt<std::is_floating_point,
#>  332 :             VecR>* = nullptr, stan::require_vector_like_vt<std::is_integral,
#>  333 :             VecI>* = nullptr, stan::require_vector_vt<std::is_floating_point,
#>  334 :             VecVar>* = nullptr>
#>  335 :   inline void
#>  336 :   write_array_impl(RNG& base_rng__, VecR& params_r__, VecI& params_i__,
#>  337 :                    VecVar& vars__, const bool
#>  338 :                    emit_transformed_parameters__ = true, const bool
#>  339 :                    emit_generated_quantities__ = true, std::ostream*
#>  340 :                    pstream__ = nullptr) const {
#>  341 :     using local_scalar_t__ = double;
#>  342 :     stan::io::deserializer<local_scalar_t__> in__(params_r__, params_i__);
#>  343 :     stan::io::serializer<local_scalar_t__> out__(vars__);
#>  344 :     static constexpr bool propto__ = true;
#>  345 :     // suppress unused var warning
#>  346 :     (void) propto__;
#>  347 :     double lp__ = 0.0;
#>  348 :     // suppress unused var warning
#>  349 :     (void) lp__;
#>  350 :     int current_statement__ = 0;
#>  351 :     stan::math::accumulator<double> lp_accum__;
#>  352 :     local_scalar_t__ DUMMY_VAR__(std::numeric_limits<double>::quiet_NaN());
#>  353 :     // suppress unused var warning
#>  354 :     (void) DUMMY_VAR__;
#>  355 :     constexpr bool jacobian__ = false;
#>  356 :     static constexpr const char* function__ =
#>  357 :       "model1bc35ea325f1__namespace::write_array";
#>  358 :     // suppress unused var warning
#>  359 :     (void) function__;
#>  360 :     try {
#>  361 :       Eigen::Matrix<double,-1,-1> w1 =
#>  362 :         Eigen::Matrix<double,-1,-1>::Constant(m, nodes,
#>  363 :           std::numeric_limits<double>::quiet_NaN());
#>  364 :       current_statement__ = 1;
#>  365 :       w1 = in__.template read<Eigen::Matrix<local_scalar_t__,-1,-1>>(m,
#>  366 :              nodes);
#>  367 :       Eigen::Matrix<double,-1,1> b1 =
#>  368 :         Eigen::Matrix<double,-1,1>::Constant(nodes,
#>  369 :           std::numeric_limits<double>::quiet_NaN());
#>  370 :       current_statement__ = 2;
#>  371 :       b1 = in__.template read<Eigen::Matrix<local_scalar_t__,-1,1>>(nodes);
#>  372 :       Eigen::Matrix<double,-1,1> w_out =
#>  373 :         Eigen::Matrix<double,-1,1>::Constant(nodes,
#>  374 :           std::numeric_limits<double>::quiet_NaN());
#>  375 :       current_statement__ = 3;
#>  376 :       w_out = in__.template read<Eigen::Matrix<local_scalar_t__,-1,1>>(nodes);
#>  377 :       double b_out = std::numeric_limits<double>::quiet_NaN();
#>  378 :       current_statement__ = 4;
#>  379 :       b_out = in__.template read<local_scalar_t__>();
#>  380 :       double sigma = std::numeric_limits<double>::quiet_NaN();
#>  381 :       current_statement__ = 5;
#>  382 :       sigma = in__.template read_constrain_lb<local_scalar_t__,
#>  383 :                 jacobian__>(0, lp__);
#>  384 :       Eigen::Matrix<double,-1,-1> z1 =
#>  385 :         Eigen::Matrix<double,-1,-1>::Constant(n, nodes,
#>  386 :           std::numeric_limits<double>::quiet_NaN());
#>  387 :       Eigen::Matrix<double,-1,-1> a1 =
#>  388 :         Eigen::Matrix<double,-1,-1>::Constant(n, nodes,
#>  389 :           std::numeric_limits<double>::quiet_NaN());
#>  390 :       Eigen::Matrix<double,-1,1> y_hat =
#>  391 :         Eigen::Matrix<double,-1,1>::Constant(n,
#>  392 :           std::numeric_limits<double>::quiet_NaN());
#>  393 :       out__.write(w1);
#>  394 :       out__.write(b1);
#>  395 :       out__.write(w_out);
#>  396 :       out__.write(b_out);
#>  397 :       out__.write(sigma);
#>  398 :       if (stan::math::logical_negation(
#>  399 :             (stan::math::primitive_value(emit_transformed_parameters__) ||
#>  400 :             stan::math::primitive_value(emit_generated_quantities__)))) {
#>  401 :         return ;
#>  402 :       }
#>  403 :       current_statement__ = 9;
#>  404 :       stan::model::assign(z1,
#>  405 :         stan::math::add(stan::math::multiply(X, w1),
#>  406 :           stan::math::rep_matrix(stan::math::transpose(b1), n)),
#>  407 :         "assigning variable z1");
#>  408 :       current_statement__ = 16;
#>  409 :       if (stan::math::logical_eq(act_fn, 1)) {
#>  410 :         current_statement__ = 15;
#>  411 :         stan::model::assign(a1, stan::math::tanh(z1), "assigning variable a1");
#>  412 :       } else {
#>  413 :         current_statement__ = 14;
#>  414 :         if (stan::math::logical_eq(act_fn, 2)) {
#>  415 :           current_statement__ = 13;
#>  416 :           stan::model::assign(a1, stan::math::inv_logit(z1),
#>  417 :             "assigning variable a1");
#>  418 :         } else {
#>  419 :           current_statement__ = 12;
#>  420 :           if (stan::math::logical_eq(act_fn, 3)) {
#>  421 :             current_statement__ = 11;
#>  422 :             stan::model::assign(a1,
#>  423 :               stan::math::log(stan::math::add(1, stan::math::exp(z1))),
#>  424 :               "assigning variable a1");
#>  425 :           } else {
#>  426 :             current_statement__ = 10;
#>  427 :             stan::model::assign(a1,
#>  428 :               stan::math::fmax(stan::math::rep_matrix(0, n, nodes), z1),
#>  429 :               "assigning variable a1");
#>  430 :           }
#>  431 :         }
#>  432 :       }
#>  433 :       current_statement__ = 17;
#>  434 :       stan::model::assign(y_hat,
#>  435 :         stan::math::add(stan::math::multiply(a1, w_out), b_out),
#>  436 :         "assigning variable y_hat");
#>  437 :       if (emit_transformed_parameters__) {
#>  438 :         out__.write(z1);
#>  439 :         out__.write(a1);
#>  440 :         out__.write(y_hat);
#>  441 :       }
#>  442 :       if (stan::math::logical_negation(emit_generated_quantities__)) {
#>  443 :         return ;
#>  444 :       }
#>  445 :     } catch (const std::exception& e) {
#>  446 :       stan::lang::rethrow_located(e, locations_array__[current_statement__]);
#>  447 :     }
#>  448 :   }
#>  449 :   template <typename VecVar, typename VecI,
#>  450 :             stan::require_vector_t<VecVar>* = nullptr,
#>  451 :             stan::require_vector_like_vt<std::is_integral, VecI>* = nullptr>
#>  452 :   inline void
#>  453 :   unconstrain_array_impl(const VecVar& params_r__, const VecI& params_i__,
#>  454 :                          VecVar& vars__, std::ostream* pstream__ = nullptr) const {
#>  455 :     using local_scalar_t__ = double;
#>  456 :     stan::io::deserializer<local_scalar_t__> in__(params_r__, params_i__);
#>  457 :     stan::io::serializer<local_scalar_t__> out__(vars__);
#>  458 :     int current_statement__ = 0;
#>  459 :     local_scalar_t__ DUMMY_VAR__(std::numeric_limits<double>::quiet_NaN());
#>  460 :     // suppress unused var warning
#>  461 :     (void) DUMMY_VAR__;
#>  462 :     try {
#>  463 :       int pos__ = std::numeric_limits<int>::min();
#>  464 :       pos__ = 1;
#>  465 :       Eigen::Matrix<local_scalar_t__,-1,-1> w1 =
#>  466 :         Eigen::Matrix<local_scalar_t__,-1,-1>::Constant(m, nodes,
#>  467 :           DUMMY_VAR__);
#>  468 :       current_statement__ = 1;
#>  469 :       stan::model::assign(w1,
#>  470 :         in__.read<Eigen::Matrix<local_scalar_t__,-1,-1>>(m, nodes),
#>  471 :         "assigning variable w1");
#>  472 :       out__.write(w1);
#>  473 :       Eigen::Matrix<local_scalar_t__,-1,1> b1 =
#>  474 :         Eigen::Matrix<local_scalar_t__,-1,1>::Constant(nodes, DUMMY_VAR__);
#>  475 :       current_statement__ = 2;
#>  476 :       stan::model::assign(b1,
#>  477 :         in__.read<Eigen::Matrix<local_scalar_t__,-1,1>>(nodes),
#>  478 :         "assigning variable b1");
#>  479 :       out__.write(b1);
#>  480 :       Eigen::Matrix<local_scalar_t__,-1,1> w_out =
#>  481 :         Eigen::Matrix<local_scalar_t__,-1,1>::Constant(nodes, DUMMY_VAR__);
#>  482 :       current_statement__ = 3;
#>  483 :       stan::model::assign(w_out,
#>  484 :         in__.read<Eigen::Matrix<local_scalar_t__,-1,1>>(nodes),
#>  485 :         "assigning variable w_out");
#>  486 :       out__.write(w_out);
#>  487 :       local_scalar_t__ b_out = DUMMY_VAR__;
#>  488 :       current_statement__ = 4;
#>  489 :       b_out = in__.read<local_scalar_t__>();
#>  490 :       out__.write(b_out);
#>  491 :       local_scalar_t__ sigma = DUMMY_VAR__;
#>  492 :       current_statement__ = 5;
#>  493 :       sigma = in__.read<local_scalar_t__>();
#>  494 :       out__.write_free_lb(0, sigma);
#>  495 :     } catch (const std::exception& e) {
#>  496 :       stan::lang::rethrow_located(e, locations_array__[current_statement__]);
#>  497 :     }
#>  498 :   }
#>  499 :   template <typename VecVar, stan::require_vector_t<VecVar>* = nullptr>
#>  500 :   inline void
#>  501 :   transform_inits_impl(const stan::io::var_context& context__, VecVar&
#>  502 :                        vars__, std::ostream* pstream__ = nullptr) const {
#>  503 :     using local_scalar_t__ = double;
#>  504 :     stan::io::serializer<local_scalar_t__> out__(vars__);
#>  505 :     int current_statement__ = 0;
#>  506 :     local_scalar_t__ DUMMY_VAR__(std::numeric_limits<double>::quiet_NaN());
#>  507 :     // suppress unused var warning
#>  508 :     (void) DUMMY_VAR__;
#>  509 :     try {
#>  510 :       current_statement__ = 1;
#>  511 :       context__.validate_dims("parameter initialization", "w1", "double",
#>  512 :         std::vector<size_t>{static_cast<size_t>(m),
#>  513 :           static_cast<size_t>(nodes)});
#>  514 :       current_statement__ = 2;
#>  515 :       context__.validate_dims("parameter initialization", "b1", "double",
#>  516 :         std::vector<size_t>{static_cast<size_t>(nodes)});
#>  517 :       current_statement__ = 3;
#>  518 :       context__.validate_dims("parameter initialization", "w_out", "double",
#>  519 :         std::vector<size_t>{static_cast<size_t>(nodes)});
#>  520 :       current_statement__ = 4;
#>  521 :       context__.validate_dims("parameter initialization", "b_out", "double",
#>  522 :         std::vector<size_t>{});
#>  523 :       current_statement__ = 5;
#>  524 :       context__.validate_dims("parameter initialization", "sigma", "double",
#>  525 :         std::vector<size_t>{});
#>  526 :       int pos__ = std::numeric_limits<int>::min();
#>  527 :       pos__ = 1;
#>  528 :       Eigen::Matrix<local_scalar_t__,-1,-1> w1 =
#>  529 :         Eigen::Matrix<local_scalar_t__,-1,-1>::Constant(m, nodes,
#>  530 :           DUMMY_VAR__);
#>  531 :       {
#>  532 :         std::vector<local_scalar_t__> w1_flat__;
#>  533 :         current_statement__ = 1;
#>  534 :         w1_flat__ = context__.vals_r("w1");
#>  535 :         current_statement__ = 1;
#>  536 :         pos__ = 1;
#>  537 :         current_statement__ = 1;
#>  538 :         for (int sym1__ = 1; sym1__ <= nodes; ++sym1__) {
#>  539 :           current_statement__ = 1;
#>  540 :           for (int sym2__ = 1; sym2__ <= m; ++sym2__) {
#>  541 :             current_statement__ = 1;
#>  542 :             stan::model::assign(w1, w1_flat__[(pos__ - 1)],
#>  543 :               "assigning variable w1", stan::model::index_uni(sym2__),
#>  544 :               stan::model::index_uni(sym1__));
#>  545 :             current_statement__ = 1;
#>  546 :             pos__ = (pos__ + 1);
#>  547 :           }
#>  548 :         }
#>  549 :       }
#>  550 :       out__.write(w1);
#>  551 :       Eigen::Matrix<local_scalar_t__,-1,1> b1 =
#>  552 :         Eigen::Matrix<local_scalar_t__,-1,1>::Constant(nodes, DUMMY_VAR__);
#>  553 :       {
#>  554 :         std::vector<local_scalar_t__> b1_flat__;
#>  555 :         current_statement__ = 2;
#>  556 :         b1_flat__ = context__.vals_r("b1");
#>  557 :         current_statement__ = 2;
#>  558 :         pos__ = 1;
#>  559 :         current_statement__ = 2;
#>  560 :         for (int sym1__ = 1; sym1__ <= nodes; ++sym1__) {
#>  561 :           current_statement__ = 2;
#>  562 :           stan::model::assign(b1, b1_flat__[(pos__ - 1)],
#>  563 :             "assigning variable b1", stan::model::index_uni(sym1__));
#>  564 :           current_statement__ = 2;
#>  565 :           pos__ = (pos__ + 1);
#>  566 :         }
#>  567 :       }
#>  568 :       out__.write(b1);
#>  569 :       Eigen::Matrix<local_scalar_t__,-1,1> w_out =
#>  570 :         Eigen::Matrix<local_scalar_t__,-1,1>::Constant(nodes, DUMMY_VAR__);
#>  571 :       {
#>  572 :         std::vector<local_scalar_t__> w_out_flat__;
#>  573 :         current_statement__ = 3;
#>  574 :         w_out_flat__ = context__.vals_r("w_out");
#>  575 :         current_statement__ = 3;
#>  576 :         pos__ = 1;
#>  577 :         current_statement__ = 3;
#>  578 :         for (int sym1__ = 1; sym1__ <= nodes; ++sym1__) {
#>  579 :           current_statement__ = 3;
#>  580 :           stan::model::assign(w_out, w_out_flat__[(pos__ - 1)],
#>  581 :             "assigning variable w_out", stan::model::index_uni(sym1__));
#>  582 :           current_statement__ = 3;
#>  583 :           pos__ = (pos__ + 1);
#>  584 :         }
#>  585 :       }
#>  586 :       out__.write(w_out);
#>  587 :       local_scalar_t__ b_out = DUMMY_VAR__;
#>  588 :       current_statement__ = 4;
#>  589 :       b_out = context__.vals_r("b_out")[(1 - 1)];
#>  590 :       out__.write(b_out);
#>  591 :       local_scalar_t__ sigma = DUMMY_VAR__;
#>  592 :       current_statement__ = 5;
#>  593 :       sigma = context__.vals_r("sigma")[(1 - 1)];
#>  594 :       out__.write_free_lb(0, sigma);
#>  595 :     } catch (const std::exception& e) {
#>  596 :       stan::lang::rethrow_located(e, locations_array__[current_statement__]);
#>  597 :     }
#>  598 :   }
#>  599 :   inline void
#>  600 :   get_param_names(std::vector<std::string>& names__, const bool
#>  601 :                   emit_transformed_parameters__ = true, const bool
#>  602 :                   emit_generated_quantities__ = true) const {
#>  603 :     names__ = std::vector<std::string>{"w1", "b1", "w_out", "b_out", "sigma"};
#>  604 :     if (emit_transformed_parameters__) {
#>  605 :       std::vector<std::string> temp{"z1", "a1", "y_hat"};
#>  606 :       names__.reserve(names__.size() + temp.size());
#>  607 :       names__.insert(names__.end(), temp.begin(), temp.end());
#>  608 :     }
#>  609 :     if (emit_generated_quantities__) {}
#>  610 :   }
#>  611 :   inline void
#>  612 :   get_dims(std::vector<std::vector<size_t>>& dimss__, const bool
#>  613 :            emit_transformed_parameters__ = true, const bool
#>  614 :            emit_generated_quantities__ = true) const {
#>  615 :     dimss__ = std::vector<std::vector<size_t>>{std::vector<size_t>{static_cast<
#>  616 :                                                                     size_t>(m),
#>  617 :                                                  static_cast<size_t>(nodes)},
#>  618 :                 std::vector<size_t>{static_cast<size_t>(nodes)},
#>  619 :                 std::vector<size_t>{static_cast<size_t>(nodes)},
#>  620 :                 std::vector<size_t>{}, std::vector<size_t>{}};
#>  621 :     if (emit_transformed_parameters__) {
#>  622 :       std::vector<std::vector<size_t>>
#>  623 :         temp{std::vector<size_t>{static_cast<size_t>(n),
#>  624 :                static_cast<size_t>(nodes)},
#>  625 :              std::vector<size_t>{static_cast<size_t>(n),
#>  626 :                static_cast<size_t>(nodes)},
#>  627 :              std::vector<size_t>{static_cast<size_t>(n)}};
#>  628 :       dimss__.reserve(dimss__.size() + temp.size());
#>  629 :       dimss__.insert(dimss__.end(), temp.begin(), temp.end());
#>  630 :     }
#>  631 :     if (emit_generated_quantities__) {}
#>  632 :   }
#>  633 :   inline void
#>  634 :   constrained_param_names(std::vector<std::string>& param_names__, bool
#>  635 :                           emit_transformed_parameters__ = true, bool
#>  636 :                           emit_generated_quantities__ = true) const final {
#>  637 :     for (int sym1__ = 1; sym1__ <= nodes; ++sym1__) {
#>  638 :       for (int sym2__ = 1; sym2__ <= m; ++sym2__) {
#>  639 :         param_names__.emplace_back(std::string() + "w1" + '.' +
#>  640 :           std::to_string(sym2__) + '.' + std::to_string(sym1__));
#>  641 :       }
#>  642 :     }
#>  643 :     for (int sym1__ = 1; sym1__ <= nodes; ++sym1__) {
#>  644 :       param_names__.emplace_back(std::string() + "b1" + '.' +
#>  645 :         std::to_string(sym1__));
#>  646 :     }
#>  647 :     for (int sym1__ = 1; sym1__ <= nodes; ++sym1__) {
#>  648 :       param_names__.emplace_back(std::string() + "w_out" + '.' +
#>  649 :         std::to_string(sym1__));
#>  650 :     }
#>  651 :     param_names__.emplace_back(std::string() + "b_out");
#>  652 :     param_names__.emplace_back(std::string() + "sigma");
#>  653 :     if (emit_transformed_parameters__) {
#>  654 :       for (int sym1__ = 1; sym1__ <= nodes; ++sym1__) {
#>  655 :         for (int sym2__ = 1; sym2__ <= n; ++sym2__) {
#>  656 :           param_names__.emplace_back(std::string() + "z1" + '.' +
#>  657 :             std::to_string(sym2__) + '.' + std::to_string(sym1__));
#>  658 :         }
#>  659 :       }
#>  660 :       for (int sym1__ = 1; sym1__ <= nodes; ++sym1__) {
#>  661 :         for (int sym2__ = 1; sym2__ <= n; ++sym2__) {
#>  662 :           param_names__.emplace_back(std::string() + "a1" + '.' +
#>  663 :             std::to_string(sym2__) + '.' + std::to_string(sym1__));
#>  664 :         }
#>  665 :       }
#>  666 :       for (int sym1__ = 1; sym1__ <= n; ++sym1__) {
#>  667 :         param_names__.emplace_back(std::string() + "y_hat" + '.' +
#>  668 :           std::to_string(sym1__));
#>  669 :       }
#>  670 :     }
#>  671 :     if (emit_generated_quantities__) {}
#>  672 :   }
#>  673 :   inline void
#>  674 :   unconstrained_param_names(std::vector<std::string>& param_names__, bool
#>  675 :                             emit_transformed_parameters__ = true, bool
#>  676 :                             emit_generated_quantities__ = true) const final {
#>  677 :     for (int sym1__ = 1; sym1__ <= nodes; ++sym1__) {
#>  678 :       for (int sym2__ = 1; sym2__ <= m; ++sym2__) {
#>  679 :         param_names__.emplace_back(std::string() + "w1" + '.' +
#>  680 :           std::to_string(sym2__) + '.' + std::to_string(sym1__));
#>  681 :       }
#>  682 :     }
#>  683 :     for (int sym1__ = 1; sym1__ <= nodes; ++sym1__) {
#>  684 :       param_names__.emplace_back(std::string() + "b1" + '.' +
#>  685 :         std::to_string(sym1__));
#>  686 :     }
#>  687 :     for (int sym1__ = 1; sym1__ <= nodes; ++sym1__) {
#>  688 :       param_names__.emplace_back(std::string() + "w_out" + '.' +
#>  689 :         std::to_string(sym1__));
#>  690 :     }
#>  691 :     param_names__.emplace_back(std::string() + "b_out");
#>  692 :     param_names__.emplace_back(std::string() + "sigma");
#>  693 :     if (emit_transformed_parameters__) {
#>  694 :       for (int sym1__ = 1; sym1__ <= nodes; ++sym1__) {
#>  695 :         for (int sym2__ = 1; sym2__ <= n; ++sym2__) {
#>  696 :           param_names__.emplace_back(std::string() + "z1" + '.' +
#>  697 :             std::to_string(sym2__) + '.' + std::to_string(sym1__));
#>  698 :         }
#>  699 :       }
#>  700 :       for (int sym1__ = 1; sym1__ <= nodes; ++sym1__) {
#>  701 :         for (int sym2__ = 1; sym2__ <= n; ++sym2__) {
#>  702 :           param_names__.emplace_back(std::string() + "a1" + '.' +
#>  703 :             std::to_string(sym2__) + '.' + std::to_string(sym1__));
#>  704 :         }
#>  705 :       }
#>  706 :       for (int sym1__ = 1; sym1__ <= n; ++sym1__) {
#>  707 :         param_names__.emplace_back(std::string() + "y_hat" + '.' +
#>  708 :           std::to_string(sym1__));
#>  709 :       }
#>  710 :     }
#>  711 :     if (emit_generated_quantities__) {}
#>  712 :   }
#>  713 :   inline std::string get_constrained_sizedtypes() const {
#>  714 :     return std::string("[{\"name\":\"w1\",\"type\":{\"name\":\"matrix\",\"rows\":" + std::to_string(m) + ",\"cols\":" + std::to_string(nodes) + "},\"block\":\"parameters\"},{\"name\":\"b1\",\"type\":{\"name\":\"vector\",\"length\":" + std::to_string(nodes) + "},\"block\":\"parameters\"},{\"name\":\"w_out\",\"type\":{\"name\":\"vector\",\"length\":" + std::to_string(nodes) + "},\"block\":\"parameters\"},{\"name\":\"b_out\",\"type\":{\"name\":\"real\"},\"block\":\"parameters\"},{\"name\":\"sigma\",\"type\":{\"name\":\"real\"},\"block\":\"parameters\"},{\"name\":\"z1\",\"type\":{\"name\":\"matrix\",\"rows\":" + std::to_string(n) + ",\"cols\":" + std::to_string(nodes) + "},\"block\":\"transformed_parameters\"},{\"name\":\"a1\",\"type\":{\"name\":\"matrix\",\"rows\":" + std::to_string(n) + ",\"cols\":" + std::to_string(nodes) + "},\"block\":\"transformed_parameters\"},{\"name\":\"y_hat\",\"type\":{\"name\":\"vector\",\"length\":" + std::to_string(n) + "},\"block\":\"transformed_parameters\"}]");
#>  715 :   }
#>  716 :   inline std::string get_unconstrained_sizedtypes() const {
#>  717 :     return std::string("[{\"name\":\"w1\",\"type\":{\"name\":\"matrix\",\"rows\":" + std::to_string(m) + ",\"cols\":" + std::to_string(nodes) + "},\"block\":\"parameters\"},{\"name\":\"b1\",\"type\":{\"name\":\"vector\",\"length\":" + std::to_string(nodes) + "},\"block\":\"parameters\"},{\"name\":\"w_out\",\"type\":{\"name\":\"vector\",\"length\":" + std::to_string(nodes) + "},\"block\":\"parameters\"},{\"name\":\"b_out\",\"type\":{\"name\":\"real\"},\"block\":\"parameters\"},{\"name\":\"sigma\",\"type\":{\"name\":\"real\"},\"block\":\"parameters\"},{\"name\":\"z1\",\"type\":{\"name\":\"matrix\",\"rows\":" + std::to_string(n) + ",\"cols\":" + std::to_string(nodes) + "},\"block\":\"transformed_parameters\"},{\"name\":\"a1\",\"type\":{\"name\":\"matrix\",\"rows\":" + std::to_string(n) + ",\"cols\":" + std::to_string(nodes) + "},\"block\":\"transformed_parameters\"},{\"name\":\"y_hat\",\"type\":{\"name\":\"vector\",\"length\":" + std::to_string(n) + "},\"block\":\"transformed_parameters\"}]");
#>  718 :   }
#>  719 :   // Begin method overload boilerplate
#>  720 :   template <typename RNG> inline void
#>  721 :   write_array(RNG& base_rng, Eigen::Matrix<double,-1,1>& params_r,
#>  722 :               Eigen::Matrix<double,-1,1>& vars, const bool
#>  723 :               emit_transformed_parameters = true, const bool
#>  724 :               emit_generated_quantities = true, std::ostream*
#>  725 :               pstream = nullptr) const {
#>  726 :     const size_t num_params__ = (((((m * nodes) + nodes) + nodes) + 1) + 1);
#>  727 :     const size_t num_transformed = emit_transformed_parameters * ((((n *
#>  728 :       nodes) + (n * nodes)) + n));
#>  729 :     const size_t num_gen_quantities = emit_generated_quantities * (0);
#>  730 :     const size_t num_to_write = num_params__ + num_transformed +
#>  731 :       num_gen_quantities;
#>  732 :     std::vector<int> params_i;
#>  733 :     vars = Eigen::Matrix<double,-1,1>::Constant(num_to_write,
#>  734 :              std::numeric_limits<double>::quiet_NaN());
#>  735 :     write_array_impl(base_rng, params_r, params_i, vars,
#>  736 :       emit_transformed_parameters, emit_generated_quantities, pstream);
#>  737 :   }
#>  738 :   template <typename RNG> inline void
#>  739 :   write_array(RNG& base_rng, std::vector<double>& params_r, std::vector<int>&
#>  740 :               params_i, std::vector<double>& vars, bool
#>  741 :               emit_transformed_parameters = true, bool
#>  742 :               emit_generated_quantities = true, std::ostream*
#>  743 :               pstream = nullptr) const {
#>  744 :     const size_t num_params__ = (((((m * nodes) + nodes) + nodes) + 1) + 1);
#>  745 :     const size_t num_transformed = emit_transformed_parameters * ((((n *
#>  746 :       nodes) + (n * nodes)) + n));
#>  747 :     const size_t num_gen_quantities = emit_generated_quantities * (0);
#>  748 :     const size_t num_to_write = num_params__ + num_transformed +
#>  749 :       num_gen_quantities;
#>  750 :     vars = std::vector<double>(num_to_write,
#>  751 :              std::numeric_limits<double>::quiet_NaN());
#>  752 :     write_array_impl(base_rng, params_r, params_i, vars,
#>  753 :       emit_transformed_parameters, emit_generated_quantities, pstream);
#>  754 :   }
#>  755 :   template <bool propto__, bool jacobian__, typename T_> inline T_
#>  756 :   log_prob(Eigen::Matrix<T_,-1,1>& params_r, std::ostream* pstream = nullptr) const {
#>  757 :     Eigen::Matrix<int,-1,1> params_i;
#>  758 :     return log_prob_impl<propto__, jacobian__>(params_r, params_i, pstream);
#>  759 :   }
#>  760 :   template <bool propto__, bool jacobian__, typename T_> inline T_
#>  761 :   log_prob(std::vector<T_>& params_r, std::vector<int>& params_i,
#>  762 :            std::ostream* pstream = nullptr) const {
#>  763 :     return log_prob_impl<propto__, jacobian__>(params_r, params_i, pstream);
#>  764 :   }
#>  765 :   inline void
#>  766 :   transform_inits(const stan::io::var_context& context,
#>  767 :                   Eigen::Matrix<double,-1,1>& params_r, std::ostream*
#>  768 :                   pstream = nullptr) const final {
#>  769 :     std::vector<double> params_r_vec(params_r.size());
#>  770 :     std::vector<int> params_i;
#>  771 :     transform_inits(context, params_i, params_r_vec, pstream);
#>  772 :     params_r = Eigen::Map<Eigen::Matrix<double,-1,1>>(params_r_vec.data(),
#>  773 :                  params_r_vec.size());
#>  774 :   }
#>  775 :   inline void
#>  776 :   transform_inits(const stan::io::var_context& context, std::vector<int>&
#>  777 :                   params_i, std::vector<double>& vars, std::ostream*
#>  778 :                   pstream__ = nullptr) const {
#>  779 :     vars.resize(num_params_r__);
#>  780 :     transform_inits_impl(context, vars, pstream__);
#>  781 :   }
#>  782 :   inline void
#>  783 :   unconstrain_array(const std::vector<double>& params_constrained,
#>  784 :                     std::vector<double>& params_unconstrained, std::ostream*
#>  785 :                     pstream = nullptr) const {
#>  786 :     const std::vector<int> params_i;
#>  787 :     params_unconstrained = std::vector<double>(num_params_r__,
#>  788 :                              std::numeric_limits<double>::quiet_NaN());
#>  789 :     unconstrain_array_impl(params_constrained, params_i,
#>  790 :       params_unconstrained, pstream);
#>  791 :   }
#>  792 :   inline void
#>  793 :   unconstrain_array(const Eigen::Matrix<double,-1,1>& params_constrained,
#>  794 :                     Eigen::Matrix<double,-1,1>& params_unconstrained,
#>  795 :                     std::ostream* pstream = nullptr) const {
#>  796 :     const std::vector<int> params_i;
#>  797 :     params_unconstrained = Eigen::Matrix<double,-1,1>::Constant(num_params_r__,
#>  798 :                              std::numeric_limits<double>::quiet_NaN());
#>  799 :     unconstrain_array_impl(params_constrained, params_i,
#>  800 :       params_unconstrained, pstream);
#>  801 :   }
#>  802 : };
#>  803 : }
#>  804 : using stan_model = model1bc35ea325f1__namespace::model1bc35ea325f1_;
#>  805 : #ifndef USING_R
#>  806 : // Boilerplate
#>  807 : stan::model::model_base&
#>  808 : new_model(stan::io::var_context& data_context, unsigned int seed,
#>  809 :           std::ostream* msg_stream) {
#>  810 :   stan_model* m = new stan_model(data_context, seed, msg_stream);
#>  811 :   return *m;
#>  812 : }
#>  813 : stan::math::profile_map& get_stan_profile_data() {
#>  814 :   return model1bc35ea325f1__namespace::profiles__;
#>  815 : }
#>  816 : #endif
#>  817 : #endif
#>  818 : 
#>  819 : RCPP_MODULE(stan_fit4model1bc35ea325f1__mod) {
#>  820 :   class_<rstan::stan_fit<stan_model, boost::random::ecuyer1988> >(
#>  821 :       "stan_fit4model1bc35ea325f1_")
#>  822 : 
#>  823 :       .constructor<SEXP, SEXP, SEXP>()
#>  824 : 
#>  825 :       .method(
#>  826 :           "call_sampler",
#>  827 :           &rstan::stan_fit<stan_model, boost::random::ecuyer1988>::call_sampler)
#>  828 :       .method(
#>  829 :           "param_names",
#>  830 :           &rstan::stan_fit<stan_model, boost::random::ecuyer1988>::param_names)
#>  831 :       .method("param_names_oi",
#>  832 :               &rstan::stan_fit<stan_model,
#>  833 :                                boost::random::ecuyer1988>::param_names_oi)
#>  834 :       .method("param_fnames_oi",
#>  835 :               &rstan::stan_fit<stan_model,
#>  836 :                                boost::random::ecuyer1988>::param_fnames_oi)
#>  837 :       .method(
#>  838 :           "param_dims",
#>  839 :           &rstan::stan_fit<stan_model, boost::random::ecuyer1988>::param_dims)
#>  840 :       .method("param_dims_oi",
#>  841 :               &rstan::stan_fit<stan_model,
#>  842 :                                boost::random::ecuyer1988>::param_dims_oi)
#>  843 :       .method("update_param_oi",
#>  844 :               &rstan::stan_fit<stan_model,
#>  845 :                                boost::random::ecuyer1988>::update_param_oi)
#>  846 :       .method("param_oi_tidx",
#>  847 :               &rstan::stan_fit<stan_model,
#>  848 :                                boost::random::ecuyer1988>::param_oi_tidx)
#>  849 :       .method("grad_log_prob",
#>  850 :               &rstan::stan_fit<stan_model,
#>  851 :                                boost::random::ecuyer1988>::grad_log_prob)
#>  852 :       .method("log_prob",
#>  853 :               &rstan::stan_fit<stan_model, boost::random::ecuyer1988>::log_prob)
#>  854 :       .method("unconstrain_pars",
#>  855 :               &rstan::stan_fit<stan_model,
#>  856 :                                boost::random::ecuyer1988>::unconstrain_pars)
#>  857 :       .method("constrain_pars",
#>  858 :               &rstan::stan_fit<stan_model,
#>  859 :                                boost::random::ecuyer1988>::constrain_pars)
#>  860 :       .method(
#>  861 :           "num_pars_unconstrained",
#>  862 :           &rstan::stan_fit<stan_model,
#>  863 :                            boost::random::ecuyer1988>::num_pars_unconstrained)
#>  864 :       .method(
#>  865 :           "unconstrained_param_names",
#>  866 :           &rstan::stan_fit<
#>  867 :               stan_model, boost::random::ecuyer1988>::unconstrained_param_names)
#>  868 :       .method(
#>  869 :           "constrained_param_names",
#>  870 :           &rstan::stan_fit<stan_model,
#>  871 :                            boost::random::ecuyer1988>::constrained_param_names)
#>  872 :       .method("standalone_gqs",
#>  873 :               &rstan::stan_fit<stan_model,
#>  874 :                                boost::random::ecuyer1988>::standalone_gqs);
#>  875 : }
#>  876 : 
#>  877 : 
#>  878 : // declarations
#>  879 : extern "C" {
#>  880 : SEXP file1bc3783593f7( ) ;
#>  881 : }
#>  882 : 
#>  883 : // definition
#>  884 : SEXP file1bc3783593f7() {
#>  885 :  return Rcpp::wrap("anon_model");
#>  886 : }
#> 
#> CHECKING DATA AND PREPROCESSING FOR MODEL 'anon_model' NOW.
#> 
#> COMPILING MODEL 'anon_model' NOW.
#> 
#> STARTING SAMPLER FOR MODEL 'anon_model' NOW.