Internal function for training the BNN

This function performs the actual fitting of the Bayesian Neural Network. It is called by the exported bnns methods and is not intended for direct use.

Usage

bnns_train(
  train_x,
  train_y,
  L = 1,
  nodes = 16,
  act_fn = 2,
  out_act_fn = 1,
  iter = 1000,
  warmup = 200,
  thin = 1,
  chains = 2,
  cores = 2,
  seed = 123,
  ...
)

Arguments

train_x

A numeric matrix representing the input features (predictors) for training. Rows correspond to observations, and columns correspond to features.

train_y

A numeric vector representing the target values for training. Its length must match the number of rows in train_x.

L

An integer specifying the number of hidden layers in the neural network. Default is 1.

nodes

An integer or vector specifying the number of nodes in each hidden layer. If a single value is provided, it is applied to all layers. Default is 16.

act_fn

An integer or vector specifying the activation function(s) for the hidden layers. Options are:

1 for tanh

2 for sigmoid (default)

3 for ReLU

4 for softplus

out_act_fn

An integer specifying the activation function for the output layer. Options are:

1 for linear (default)

2 for sigmoid

3 for softmax

iter

An integer specifying the total number of iterations for the Stan sampler. Default is 1e3.

warmup

An integer specifying the number of warmup iterations for the Stan sampler. Default is 2e2.

thin

An integer specifying the thinning interval for Stan samples. Default is 1.

chains

An integer specifying the number of Markov chains. Default is 2.

cores

An integer specifying the number of CPU cores to use for parallel sampling. Default is 2.

seed

An integer specifying the random seed for reproducibility. Default is 123.

...

Additional arguments passed to the Stan sampler.

Value

An object of class "bnns" containing the following components:

fit

The fitted Stan model object.

call

The matched call.

data

A list containing the Stan data used in the model.

Details

The function uses the generate_stan_code function to dynamically generate Stan code based on the specified number of layers and nodes. Stan is then used to fit the Bayesian Neural Network.

See also

Examples

# Example usage:
train_x <- matrix(runif(20), nrow = 10, ncol = 2)
train_y <- rnorm(10)
model <- bnns:::bnns_train(train_x, train_y, L = 1, nodes = 2, act_fn = 2,
iter = 1e2, warmup = 5e1, chains = 1)
#> 
#> TRANSLATING MODEL '' FROM Stan CODE TO C++ CODE NOW.
#> OS: x86_64, linux-gnu; rstan: 2.32.6; Rcpp: 1.0.13.1; inline: 0.3.20 
#>  >> setting environment variables: 
#> PKG_LIBS =  '/home/runner/work/_temp/Library/rstan/lib//libStanServices.a' -L'/home/runner/work/_temp/Library/StanHeaders/lib/' -lStanHeaders -L'/home/runner/work/_temp/Library/RcppParallel/lib/' -ltbb 
#> PKG_CPPFLAGS =   -I"/home/runner/work/_temp/Library/Rcpp/include/"  -I"/home/runner/work/_temp/Library/RcppEigen/include/"  -I"/home/runner/work/_temp/Library/RcppEigen/include/unsupported"  -I"/home/runner/work/_temp/Library/BH/include" -I"/home/runner/work/_temp/Library/StanHeaders/include/src/"  -I"/home/runner/work/_temp/Library/StanHeaders/include/"  -I"/home/runner/work/_temp/Library/RcppParallel/include/"  -I"/home/runner/work/_temp/Library/rstan/include" -DEIGEN_NO_DEBUG  -DBOOST_DISABLE_ASSERTS  -DBOOST_PENDING_INTEGER_LOG2_HPP  -DSTAN_THREADS  -DUSE_STANC3 -DSTRICT_R_HEADERS  -DBOOST_PHOENIX_NO_VARIADIC_EXPRESSION  -D_HAS_AUTO_PTR_ETC=0  -include '/home/runner/work/_temp/Library/StanHeaders/include/stan/math/prim/fun/Eigen.hpp'  -D_REENTRANT -DRCPP_PARALLEL_USE_TBB=1 
#>  >> Program source :
#> 
#>    1 : 
#>    2 : // includes from the plugin
#>    3 : // [[Rcpp::plugins(cpp14)]]
#>    4 : 
#>    5 : 
#>    6 : // user includes
#>    7 : #include <Rcpp.h>
#>    8 : using namespace Rcpp;
#>    9 : #ifndef MODELS_HPP
#>   10 : #define MODELS_HPP
#>   11 : #define STAN__SERVICES__COMMAND_HPP
#>   12 : #include <rstan/rstaninc.hpp>
#>   13 : #ifndef USE_STANC3
#>   14 : #define USE_STANC3
#>   15 : #endif
#>   16 : // Code generated by stanc v2.32.2
#>   17 : #include <stan/model/model_header.hpp>
#>   18 : namespace model1bee24d635d2__namespace {
#>   19 : using stan::model::model_base_crtp;
#>   20 : using namespace stan::math;
#>   21 : stan::math::profile_map profiles__;
#>   22 : static constexpr std::array<const char*, 43> locations_array__ =
#>   23 :   {" (found before start of program)",
#>   24 :   " (in 'anon_model', line 11, column 2 to column 22)",
#>   25 :   " (in 'anon_model', line 12, column 2 to column 19)",
#>   26 :   " (in 'anon_model', line 13, column 2 to column 22)",
#>   27 :   " (in 'anon_model', line 14, column 2 to column 13)",
#>   28 :   " (in 'anon_model', line 15, column 2 to column 22)",
#>   29 :   " (in 'anon_model', line 18, column 2 to column 22)",
#>   30 :   " (in 'anon_model', line 19, column 2 to column 22)",
#>   31 :   " (in 'anon_model', line 20, column 2 to column 18)",
#>   32 :   " (in 'anon_model', line 21, column 2 to column 35)",
#>   33 :   " (in 'anon_model', line 25, column 7 to column 46)",
#>   34 :   " (in 'anon_model', line 24, column 24 to column 46)",
#>   35 :   " (in 'anon_model', line 24, column 7 to line 25, column 46)",
#>   36 :   " (in 'anon_model', line 23, column 24 to column 43)",
#>   37 :   " (in 'anon_model', line 23, column 7 to line 25, column 46)",
#>   38 :   " (in 'anon_model', line 22, column 19 to column 33)",
#>   39 :   " (in 'anon_model', line 22, column 2 to line 25, column 46)",
#>   40 :   " (in 'anon_model', line 26, column 2 to column 29)",
#>   41 :   " (in 'anon_model', line 29, column 2 to column 31)",
#>   42 :   " (in 'anon_model', line 30, column 2 to column 20)",
#>   43 :   " (in 'anon_model', line 31, column 2 to column 23)",
#>   44 :   " (in 'anon_model', line 32, column 2 to column 23)",
#>   45 :   " (in 'anon_model', line 33, column 2 to column 23)",
#>   46 :   " (in 'anon_model', line 34, column 2 to column 27)",
#>   47 :   " (in 'anon_model', line 2, column 2 to column 17)",
#>   48 :   " (in 'anon_model', line 3, column 2 to column 17)",
#>   49 :   " (in 'anon_model', line 4, column 2 to column 17)",
#>   50 :   " (in 'anon_model', line 5, column 2 to column 21)",
#>   51 :   " (in 'anon_model', line 6, column 9 to column 10)",
#>   52 :   " (in 'anon_model', line 6, column 12 to column 13)",
#>   53 :   " (in 'anon_model', line 6, column 2 to column 17)",
#>   54 :   " (in 'anon_model', line 7, column 9 to column 10)",
#>   55 :   " (in 'anon_model', line 7, column 2 to column 14)",
#>   56 :   " (in 'anon_model', line 8, column 2 to column 31)",
#>   57 :   " (in 'anon_model', line 11, column 9 to column 10)",
#>   58 :   " (in 'anon_model', line 11, column 12 to column 17)",
#>   59 :   " (in 'anon_model', line 12, column 9 to column 14)",
#>   60 :   " (in 'anon_model', line 13, column 9 to column 14)",
#>   61 :   " (in 'anon_model', line 18, column 9 to column 10)",
#>   62 :   " (in 'anon_model', line 18, column 12 to column 17)",
#>   63 :   " (in 'anon_model', line 19, column 9 to column 10)",
#>   64 :   " (in 'anon_model', line 19, column 12 to column 17)",
#>   65 :   " (in 'anon_model', line 20, column 9 to column 10)"};
#>   66 : class model1bee24d635d2_ final : public model_base_crtp<model1bee24d635d2_> {
#>   67 : private:
#>   68 :   int n;
#>   69 :   int m;
#>   70 :   int L;
#>   71 :   int nodes;
#>   72 :   Eigen::Matrix<double,-1,-1> X_data__;
#>   73 :   Eigen::Matrix<double,-1,1> y_data__;
#>   74 :   int act_fn;
#>   75 :   Eigen::Map<Eigen::Matrix<double,-1,-1>> X{nullptr, 0, 0};
#>   76 :   Eigen::Map<Eigen::Matrix<double,-1,1>> y{nullptr, 0};
#>   77 : public:
#>   78 :   ~model1bee24d635d2_() {}
#>   79 :   model1bee24d635d2_(stan::io::var_context& context__, unsigned int
#>   80 :                      random_seed__ = 0, std::ostream* pstream__ = nullptr)
#>   81 :       : model_base_crtp(0) {
#>   82 :     int current_statement__ = 0;
#>   83 :     using local_scalar_t__ = double;
#>   84 :     boost::ecuyer1988 base_rng__ =
#>   85 :       stan::services::util::create_rng(random_seed__, 0);
#>   86 :     // suppress unused var warning
#>   87 :     (void) base_rng__;
#>   88 :     static constexpr const char* function__ =
#>   89 :       "model1bee24d635d2__namespace::model1bee24d635d2_";
#>   90 :     // suppress unused var warning
#>   91 :     (void) function__;
#>   92 :     local_scalar_t__ DUMMY_VAR__(std::numeric_limits<double>::quiet_NaN());
#>   93 :     // suppress unused var warning
#>   94 :     (void) DUMMY_VAR__;
#>   95 :     try {
#>   96 :       int pos__ = std::numeric_limits<int>::min();
#>   97 :       pos__ = 1;
#>   98 :       current_statement__ = 24;
#>   99 :       context__.validate_dims("data initialization", "n", "int",
#>  100 :         std::vector<size_t>{});
#>  101 :       n = std::numeric_limits<int>::min();
#>  102 :       current_statement__ = 24;
#>  103 :       n = context__.vals_i("n")[(1 - 1)];
#>  104 :       current_statement__ = 24;
#>  105 :       stan::math::check_greater_or_equal(function__, "n", n, 1);
#>  106 :       current_statement__ = 25;
#>  107 :       context__.validate_dims("data initialization", "m", "int",
#>  108 :         std::vector<size_t>{});
#>  109 :       m = std::numeric_limits<int>::min();
#>  110 :       current_statement__ = 25;
#>  111 :       m = context__.vals_i("m")[(1 - 1)];
#>  112 :       current_statement__ = 25;
#>  113 :       stan::math::check_greater_or_equal(function__, "m", m, 1);
#>  114 :       current_statement__ = 26;
#>  115 :       context__.validate_dims("data initialization", "L", "int",
#>  116 :         std::vector<size_t>{});
#>  117 :       L = std::numeric_limits<int>::min();
#>  118 :       current_statement__ = 26;
#>  119 :       L = context__.vals_i("L")[(1 - 1)];
#>  120 :       current_statement__ = 26;
#>  121 :       stan::math::check_greater_or_equal(function__, "L", L, 1);
#>  122 :       current_statement__ = 27;
#>  123 :       context__.validate_dims("data initialization", "nodes", "int",
#>  124 :         std::vector<size_t>{});
#>  125 :       nodes = std::numeric_limits<int>::min();
#>  126 :       current_statement__ = 27;
#>  127 :       nodes = context__.vals_i("nodes")[(1 - 1)];
#>  128 :       current_statement__ = 27;
#>  129 :       stan::math::check_greater_or_equal(function__, "nodes", nodes, 1);
#>  130 :       current_statement__ = 28;
#>  131 :       stan::math::validate_non_negative_index("X", "n", n);
#>  132 :       current_statement__ = 29;
#>  133 :       stan::math::validate_non_negative_index("X", "m", m);
#>  134 :       current_statement__ = 30;
#>  135 :       context__.validate_dims("data initialization", "X", "double",
#>  136 :         std::vector<size_t>{static_cast<size_t>(n), static_cast<size_t>(m)});
#>  137 :       X_data__ = Eigen::Matrix<double,-1,-1>::Constant(n, m,
#>  138 :                    std::numeric_limits<double>::quiet_NaN());
#>  139 :       new (&X) Eigen::Map<Eigen::Matrix<double,-1,-1>>(X_data__.data(), n, m);
#>  140 :       {
#>  141 :         std::vector<local_scalar_t__> X_flat__;
#>  142 :         current_statement__ = 30;
#>  143 :         X_flat__ = context__.vals_r("X");
#>  144 :         current_statement__ = 30;
#>  145 :         pos__ = 1;
#>  146 :         current_statement__ = 30;
#>  147 :         for (int sym1__ = 1; sym1__ <= m; ++sym1__) {
#>  148 :           current_statement__ = 30;
#>  149 :           for (int sym2__ = 1; sym2__ <= n; ++sym2__) {
#>  150 :             current_statement__ = 30;
#>  151 :             stan::model::assign(X, X_flat__[(pos__ - 1)],
#>  152 :               "assigning variable X", stan::model::index_uni(sym2__),
#>  153 :               stan::model::index_uni(sym1__));
#>  154 :             current_statement__ = 30;
#>  155 :             pos__ = (pos__ + 1);
#>  156 :           }
#>  157 :         }
#>  158 :       }
#>  159 :       current_statement__ = 31;
#>  160 :       stan::math::validate_non_negative_index("y", "n", n);
#>  161 :       current_statement__ = 32;
#>  162 :       context__.validate_dims("data initialization", "y", "double",
#>  163 :         std::vector<size_t>{static_cast<size_t>(n)});
#>  164 :       y_data__ = Eigen::Matrix<double,-1,1>::Constant(n,
#>  165 :                    std::numeric_limits<double>::quiet_NaN());
#>  166 :       new (&y) Eigen::Map<Eigen::Matrix<double,-1,1>>(y_data__.data(), n);
#>  167 :       {
#>  168 :         std::vector<local_scalar_t__> y_flat__;
#>  169 :         current_statement__ = 32;
#>  170 :         y_flat__ = context__.vals_r("y");
#>  171 :         current_statement__ = 32;
#>  172 :         pos__ = 1;
#>  173 :         current_statement__ = 32;
#>  174 :         for (int sym1__ = 1; sym1__ <= n; ++sym1__) {
#>  175 :           current_statement__ = 32;
#>  176 :           stan::model::assign(y, y_flat__[(pos__ - 1)],
#>  177 :             "assigning variable y", stan::model::index_uni(sym1__));
#>  178 :           current_statement__ = 32;
#>  179 :           pos__ = (pos__ + 1);
#>  180 :         }
#>  181 :       }
#>  182 :       current_statement__ = 33;
#>  183 :       context__.validate_dims("data initialization", "act_fn", "int",
#>  184 :         std::vector<size_t>{});
#>  185 :       act_fn = std::numeric_limits<int>::min();
#>  186 :       current_statement__ = 33;
#>  187 :       act_fn = context__.vals_i("act_fn")[(1 - 1)];
#>  188 :       current_statement__ = 33;
#>  189 :       stan::math::check_greater_or_equal(function__, "act_fn", act_fn, 1);
#>  190 :       current_statement__ = 33;
#>  191 :       stan::math::check_less_or_equal(function__, "act_fn", act_fn, 3);
#>  192 :       current_statement__ = 34;
#>  193 :       stan::math::validate_non_negative_index("w1", "m", m);
#>  194 :       current_statement__ = 35;
#>  195 :       stan::math::validate_non_negative_index("w1", "nodes", nodes);
#>  196 :       current_statement__ = 36;
#>  197 :       stan::math::validate_non_negative_index("b1", "nodes", nodes);
#>  198 :       current_statement__ = 37;
#>  199 :       stan::math::validate_non_negative_index("w_out", "nodes", nodes);
#>  200 :       current_statement__ = 38;
#>  201 :       stan::math::validate_non_negative_index("z1", "n", n);
#>  202 :       current_statement__ = 39;
#>  203 :       stan::math::validate_non_negative_index("z1", "nodes", nodes);
#>  204 :       current_statement__ = 40;
#>  205 :       stan::math::validate_non_negative_index("a1", "n", n);
#>  206 :       current_statement__ = 41;
#>  207 :       stan::math::validate_non_negative_index("a1", "nodes", nodes);
#>  208 :       current_statement__ = 42;
#>  209 :       stan::math::validate_non_negative_index("y_hat", "n", n);
#>  210 :     } catch (const std::exception& e) {
#>  211 :       stan::lang::rethrow_located(e, locations_array__[current_statement__]);
#>  212 :     }
#>  213 :     num_params_r__ = (m * nodes) + nodes + nodes + 1 + 1;
#>  214 :   }
#>  215 :   inline std::string model_name() const final {
#>  216 :     return "model1bee24d635d2_";
#>  217 :   }
#>  218 :   inline std::vector<std::string> model_compile_info() const noexcept {
#>  219 :     return std::vector<std::string>{"stanc_version = stanc3 v2.32.2",
#>  220 :              "stancflags = --"};
#>  221 :   }
#>  222 :   template <bool propto__, bool jacobian__, typename VecR, typename VecI,
#>  223 :             stan::require_vector_like_t<VecR>* = nullptr,
#>  224 :             stan::require_vector_like_vt<std::is_integral, VecI>* = nullptr>
#>  225 :   inline stan::scalar_type_t<VecR>
#>  226 :   log_prob_impl(VecR& params_r__, VecI& params_i__, std::ostream*
#>  227 :                 pstream__ = nullptr) const {
#>  228 :     using T__ = stan::scalar_type_t<VecR>;
#>  229 :     using local_scalar_t__ = T__;
#>  230 :     T__ lp__(0.0);
#>  231 :     stan::math::accumulator<T__> lp_accum__;
#>  232 :     stan::io::deserializer<local_scalar_t__> in__(params_r__, params_i__);
#>  233 :     int current_statement__ = 0;
#>  234 :     local_scalar_t__ DUMMY_VAR__(std::numeric_limits<double>::quiet_NaN());
#>  235 :     // suppress unused var warning
#>  236 :     (void) DUMMY_VAR__;
#>  237 :     static constexpr const char* function__ =
#>  238 :       "model1bee24d635d2__namespace::log_prob";
#>  239 :     // suppress unused var warning
#>  240 :     (void) function__;
#>  241 :     try {
#>  242 :       Eigen::Matrix<local_scalar_t__,-1,-1> w1 =
#>  243 :         Eigen::Matrix<local_scalar_t__,-1,-1>::Constant(m, nodes,
#>  244 :           DUMMY_VAR__);
#>  245 :       current_statement__ = 1;
#>  246 :       w1 = in__.template read<Eigen::Matrix<local_scalar_t__,-1,-1>>(m,
#>  247 :              nodes);
#>  248 :       Eigen::Matrix<local_scalar_t__,-1,1> b1 =
#>  249 :         Eigen::Matrix<local_scalar_t__,-1,1>::Constant(nodes, DUMMY_VAR__);
#>  250 :       current_statement__ = 2;
#>  251 :       b1 = in__.template read<Eigen::Matrix<local_scalar_t__,-1,1>>(nodes);
#>  252 :       Eigen::Matrix<local_scalar_t__,-1,1> w_out =
#>  253 :         Eigen::Matrix<local_scalar_t__,-1,1>::Constant(nodes, DUMMY_VAR__);
#>  254 :       current_statement__ = 3;
#>  255 :       w_out = in__.template read<Eigen::Matrix<local_scalar_t__,-1,1>>(nodes);
#>  256 :       local_scalar_t__ b_out = DUMMY_VAR__;
#>  257 :       current_statement__ = 4;
#>  258 :       b_out = in__.template read<local_scalar_t__>();
#>  259 :       local_scalar_t__ sigma = DUMMY_VAR__;
#>  260 :       current_statement__ = 5;
#>  261 :       sigma = in__.template read_constrain_lb<local_scalar_t__,
#>  262 :                 jacobian__>(0, lp__);
#>  263 :       Eigen::Matrix<local_scalar_t__,-1,-1> z1 =
#>  264 :         Eigen::Matrix<local_scalar_t__,-1,-1>::Constant(n, nodes,
#>  265 :           DUMMY_VAR__);
#>  266 :       Eigen::Matrix<local_scalar_t__,-1,-1> a1 =
#>  267 :         Eigen::Matrix<local_scalar_t__,-1,-1>::Constant(n, nodes,
#>  268 :           DUMMY_VAR__);
#>  269 :       Eigen::Matrix<local_scalar_t__,-1,1> y_hat =
#>  270 :         Eigen::Matrix<local_scalar_t__,-1,1>::Constant(n, DUMMY_VAR__);
#>  271 :       current_statement__ = 9;
#>  272 :       stan::model::assign(z1,
#>  273 :         stan::math::add(stan::math::multiply(X, w1),
#>  274 :           stan::math::rep_matrix(stan::math::transpose(b1), n)),
#>  275 :         "assigning variable z1");
#>  276 :       current_statement__ = 16;
#>  277 :       if (stan::math::logical_eq(act_fn, 1)) {
#>  278 :         current_statement__ = 15;
#>  279 :         stan::model::assign(a1, stan::math::tanh(z1), "assigning variable a1");
#>  280 :       } else {
#>  281 :         current_statement__ = 14;
#>  282 :         if (stan::math::logical_eq(act_fn, 2)) {
#>  283 :           current_statement__ = 13;
#>  284 :           stan::model::assign(a1, stan::math::inv_logit(z1),
#>  285 :             "assigning variable a1");
#>  286 :         } else {
#>  287 :           current_statement__ = 12;
#>  288 :           if (stan::math::logical_eq(act_fn, 3)) {
#>  289 :             current_statement__ = 11;
#>  290 :             stan::model::assign(a1,
#>  291 :               stan::math::log(stan::math::add(1, stan::math::exp(z1))),
#>  292 :               "assigning variable a1");
#>  293 :           } else {
#>  294 :             current_statement__ = 10;
#>  295 :             stan::model::assign(a1,
#>  296 :               stan::math::fmax(stan::math::rep_matrix(0, n, nodes), z1),
#>  297 :               "assigning variable a1");
#>  298 :           }
#>  299 :         }
#>  300 :       }
#>  301 :       current_statement__ = 17;
#>  302 :       stan::model::assign(y_hat,
#>  303 :         stan::math::add(stan::math::multiply(a1, w_out), b_out),
#>  304 :         "assigning variable y_hat");
#>  305 :       {
#>  306 :         current_statement__ = 18;
#>  307 :         lp_accum__.add(stan::math::normal_lpdf<propto__>(
#>  308 :                          stan::math::to_vector(w1), 0, 1));
#>  309 :         current_statement__ = 19;
#>  310 :         lp_accum__.add(stan::math::normal_lpdf<propto__>(b1, 0, 1));
#>  311 :         current_statement__ = 20;
#>  312 :         lp_accum__.add(stan::math::normal_lpdf<propto__>(w_out, 0, 1));
#>  313 :         current_statement__ = 21;
#>  314 :         lp_accum__.add(stan::math::normal_lpdf<propto__>(b_out, 0, 1));
#>  315 :         current_statement__ = 22;
#>  316 :         lp_accum__.add(stan::math::normal_lpdf<propto__>(sigma, 0, 1));
#>  317 :         current_statement__ = 23;
#>  318 :         lp_accum__.add(stan::math::normal_lpdf<propto__>(y, y_hat, sigma));
#>  319 :       }
#>  320 :     } catch (const std::exception& e) {
#>  321 :       stan::lang::rethrow_located(e, locations_array__[current_statement__]);
#>  322 :     }
#>  323 :     lp_accum__.add(lp__);
#>  324 :     return lp_accum__.sum();
#>  325 :   }
#>  326 :   template <typename RNG, typename VecR, typename VecI, typename VecVar,
#>  327 :             stan::require_vector_like_vt<std::is_floating_point,
#>  328 :             VecR>* = nullptr, stan::require_vector_like_vt<std::is_integral,
#>  329 :             VecI>* = nullptr, stan::require_vector_vt<std::is_floating_point,
#>  330 :             VecVar>* = nullptr>
#>  331 :   inline void
#>  332 :   write_array_impl(RNG& base_rng__, VecR& params_r__, VecI& params_i__,
#>  333 :                    VecVar& vars__, const bool
#>  334 :                    emit_transformed_parameters__ = true, const bool
#>  335 :                    emit_generated_quantities__ = true, std::ostream*
#>  336 :                    pstream__ = nullptr) const {
#>  337 :     using local_scalar_t__ = double;
#>  338 :     stan::io::deserializer<local_scalar_t__> in__(params_r__, params_i__);
#>  339 :     stan::io::serializer<local_scalar_t__> out__(vars__);
#>  340 :     static constexpr bool propto__ = true;
#>  341 :     // suppress unused var warning
#>  342 :     (void) propto__;
#>  343 :     double lp__ = 0.0;
#>  344 :     // suppress unused var warning
#>  345 :     (void) lp__;
#>  346 :     int current_statement__ = 0;
#>  347 :     stan::math::accumulator<double> lp_accum__;
#>  348 :     local_scalar_t__ DUMMY_VAR__(std::numeric_limits<double>::quiet_NaN());
#>  349 :     // suppress unused var warning
#>  350 :     (void) DUMMY_VAR__;
#>  351 :     constexpr bool jacobian__ = false;
#>  352 :     static constexpr const char* function__ =
#>  353 :       "model1bee24d635d2__namespace::write_array";
#>  354 :     // suppress unused var warning
#>  355 :     (void) function__;
#>  356 :     try {
#>  357 :       Eigen::Matrix<double,-1,-1> w1 =
#>  358 :         Eigen::Matrix<double,-1,-1>::Constant(m, nodes,
#>  359 :           std::numeric_limits<double>::quiet_NaN());
#>  360 :       current_statement__ = 1;
#>  361 :       w1 = in__.template read<Eigen::Matrix<local_scalar_t__,-1,-1>>(m,
#>  362 :              nodes);
#>  363 :       Eigen::Matrix<double,-1,1> b1 =
#>  364 :         Eigen::Matrix<double,-1,1>::Constant(nodes,
#>  365 :           std::numeric_limits<double>::quiet_NaN());
#>  366 :       current_statement__ = 2;
#>  367 :       b1 = in__.template read<Eigen::Matrix<local_scalar_t__,-1,1>>(nodes);
#>  368 :       Eigen::Matrix<double,-1,1> w_out =
#>  369 :         Eigen::Matrix<double,-1,1>::Constant(nodes,
#>  370 :           std::numeric_limits<double>::quiet_NaN());
#>  371 :       current_statement__ = 3;
#>  372 :       w_out = in__.template read<Eigen::Matrix<local_scalar_t__,-1,1>>(nodes);
#>  373 :       double b_out = std::numeric_limits<double>::quiet_NaN();
#>  374 :       current_statement__ = 4;
#>  375 :       b_out = in__.template read<local_scalar_t__>();
#>  376 :       double sigma = std::numeric_limits<double>::quiet_NaN();
#>  377 :       current_statement__ = 5;
#>  378 :       sigma = in__.template read_constrain_lb<local_scalar_t__,
#>  379 :                 jacobian__>(0, lp__);
#>  380 :       Eigen::Matrix<double,-1,-1> z1 =
#>  381 :         Eigen::Matrix<double,-1,-1>::Constant(n, nodes,
#>  382 :           std::numeric_limits<double>::quiet_NaN());
#>  383 :       Eigen::Matrix<double,-1,-1> a1 =
#>  384 :         Eigen::Matrix<double,-1,-1>::Constant(n, nodes,
#>  385 :           std::numeric_limits<double>::quiet_NaN());
#>  386 :       Eigen::Matrix<double,-1,1> y_hat =
#>  387 :         Eigen::Matrix<double,-1,1>::Constant(n,
#>  388 :           std::numeric_limits<double>::quiet_NaN());
#>  389 :       out__.write(w1);
#>  390 :       out__.write(b1);
#>  391 :       out__.write(w_out);
#>  392 :       out__.write(b_out);
#>  393 :       out__.write(sigma);
#>  394 :       if (stan::math::logical_negation(
#>  395 :             (stan::math::primitive_value(emit_transformed_parameters__) ||
#>  396 :             stan::math::primitive_value(emit_generated_quantities__)))) {
#>  397 :         return ;
#>  398 :       }
#>  399 :       current_statement__ = 9;
#>  400 :       stan::model::assign(z1,
#>  401 :         stan::math::add(stan::math::multiply(X, w1),
#>  402 :           stan::math::rep_matrix(stan::math::transpose(b1), n)),
#>  403 :         "assigning variable z1");
#>  404 :       current_statement__ = 16;
#>  405 :       if (stan::math::logical_eq(act_fn, 1)) {
#>  406 :         current_statement__ = 15;
#>  407 :         stan::model::assign(a1, stan::math::tanh(z1), "assigning variable a1");
#>  408 :       } else {
#>  409 :         current_statement__ = 14;
#>  410 :         if (stan::math::logical_eq(act_fn, 2)) {
#>  411 :           current_statement__ = 13;
#>  412 :           stan::model::assign(a1, stan::math::inv_logit(z1),
#>  413 :             "assigning variable a1");
#>  414 :         } else {
#>  415 :           current_statement__ = 12;
#>  416 :           if (stan::math::logical_eq(act_fn, 3)) {
#>  417 :             current_statement__ = 11;
#>  418 :             stan::model::assign(a1,
#>  419 :               stan::math::log(stan::math::add(1, stan::math::exp(z1))),
#>  420 :               "assigning variable a1");
#>  421 :           } else {
#>  422 :             current_statement__ = 10;
#>  423 :             stan::model::assign(a1,
#>  424 :               stan::math::fmax(stan::math::rep_matrix(0, n, nodes), z1),
#>  425 :               "assigning variable a1");
#>  426 :           }
#>  427 :         }
#>  428 :       }
#>  429 :       current_statement__ = 17;
#>  430 :       stan::model::assign(y_hat,
#>  431 :         stan::math::add(stan::math::multiply(a1, w_out), b_out),
#>  432 :         "assigning variable y_hat");
#>  433 :       if (emit_transformed_parameters__) {
#>  434 :         out__.write(z1);
#>  435 :         out__.write(a1);
#>  436 :         out__.write(y_hat);
#>  437 :       }
#>  438 :       if (stan::math::logical_negation(emit_generated_quantities__)) {
#>  439 :         return ;
#>  440 :       }
#>  441 :     } catch (const std::exception& e) {
#>  442 :       stan::lang::rethrow_located(e, locations_array__[current_statement__]);
#>  443 :     }
#>  444 :   }
#>  445 :   template <typename VecVar, typename VecI,
#>  446 :             stan::require_vector_t<VecVar>* = nullptr,
#>  447 :             stan::require_vector_like_vt<std::is_integral, VecI>* = nullptr>
#>  448 :   inline void
#>  449 :   unconstrain_array_impl(const VecVar& params_r__, const VecI& params_i__,
#>  450 :                          VecVar& vars__, std::ostream* pstream__ = nullptr) const {
#>  451 :     using local_scalar_t__ = double;
#>  452 :     stan::io::deserializer<local_scalar_t__> in__(params_r__, params_i__);
#>  453 :     stan::io::serializer<local_scalar_t__> out__(vars__);
#>  454 :     int current_statement__ = 0;
#>  455 :     local_scalar_t__ DUMMY_VAR__(std::numeric_limits<double>::quiet_NaN());
#>  456 :     // suppress unused var warning
#>  457 :     (void) DUMMY_VAR__;
#>  458 :     try {
#>  459 :       int pos__ = std::numeric_limits<int>::min();
#>  460 :       pos__ = 1;
#>  461 :       Eigen::Matrix<local_scalar_t__,-1,-1> w1 =
#>  462 :         Eigen::Matrix<local_scalar_t__,-1,-1>::Constant(m, nodes,
#>  463 :           DUMMY_VAR__);
#>  464 :       current_statement__ = 1;
#>  465 :       stan::model::assign(w1,
#>  466 :         in__.read<Eigen::Matrix<local_scalar_t__,-1,-1>>(m, nodes),
#>  467 :         "assigning variable w1");
#>  468 :       out__.write(w1);
#>  469 :       Eigen::Matrix<local_scalar_t__,-1,1> b1 =
#>  470 :         Eigen::Matrix<local_scalar_t__,-1,1>::Constant(nodes, DUMMY_VAR__);
#>  471 :       current_statement__ = 2;
#>  472 :       stan::model::assign(b1,
#>  473 :         in__.read<Eigen::Matrix<local_scalar_t__,-1,1>>(nodes),
#>  474 :         "assigning variable b1");
#>  475 :       out__.write(b1);
#>  476 :       Eigen::Matrix<local_scalar_t__,-1,1> w_out =
#>  477 :         Eigen::Matrix<local_scalar_t__,-1,1>::Constant(nodes, DUMMY_VAR__);
#>  478 :       current_statement__ = 3;
#>  479 :       stan::model::assign(w_out,
#>  480 :         in__.read<Eigen::Matrix<local_scalar_t__,-1,1>>(nodes),
#>  481 :         "assigning variable w_out");
#>  482 :       out__.write(w_out);
#>  483 :       local_scalar_t__ b_out = DUMMY_VAR__;
#>  484 :       current_statement__ = 4;
#>  485 :       b_out = in__.read<local_scalar_t__>();
#>  486 :       out__.write(b_out);
#>  487 :       local_scalar_t__ sigma = DUMMY_VAR__;
#>  488 :       current_statement__ = 5;
#>  489 :       sigma = in__.read<local_scalar_t__>();
#>  490 :       out__.write_free_lb(0, sigma);
#>  491 :     } catch (const std::exception& e) {
#>  492 :       stan::lang::rethrow_located(e, locations_array__[current_statement__]);
#>  493 :     }
#>  494 :   }
#>  495 :   template <typename VecVar, stan::require_vector_t<VecVar>* = nullptr>
#>  496 :   inline void
#>  497 :   transform_inits_impl(const stan::io::var_context& context__, VecVar&
#>  498 :                        vars__, std::ostream* pstream__ = nullptr) const {
#>  499 :     using local_scalar_t__ = double;
#>  500 :     stan::io::serializer<local_scalar_t__> out__(vars__);
#>  501 :     int current_statement__ = 0;
#>  502 :     local_scalar_t__ DUMMY_VAR__(std::numeric_limits<double>::quiet_NaN());
#>  503 :     // suppress unused var warning
#>  504 :     (void) DUMMY_VAR__;
#>  505 :     try {
#>  506 :       current_statement__ = 1;
#>  507 :       context__.validate_dims("parameter initialization", "w1", "double",
#>  508 :         std::vector<size_t>{static_cast<size_t>(m),
#>  509 :           static_cast<size_t>(nodes)});
#>  510 :       current_statement__ = 2;
#>  511 :       context__.validate_dims("parameter initialization", "b1", "double",
#>  512 :         std::vector<size_t>{static_cast<size_t>(nodes)});
#>  513 :       current_statement__ = 3;
#>  514 :       context__.validate_dims("parameter initialization", "w_out", "double",
#>  515 :         std::vector<size_t>{static_cast<size_t>(nodes)});
#>  516 :       current_statement__ = 4;
#>  517 :       context__.validate_dims("parameter initialization", "b_out", "double",
#>  518 :         std::vector<size_t>{});
#>  519 :       current_statement__ = 5;
#>  520 :       context__.validate_dims("parameter initialization", "sigma", "double",
#>  521 :         std::vector<size_t>{});
#>  522 :       int pos__ = std::numeric_limits<int>::min();
#>  523 :       pos__ = 1;
#>  524 :       Eigen::Matrix<local_scalar_t__,-1,-1> w1 =
#>  525 :         Eigen::Matrix<local_scalar_t__,-1,-1>::Constant(m, nodes,
#>  526 :           DUMMY_VAR__);
#>  527 :       {
#>  528 :         std::vector<local_scalar_t__> w1_flat__;
#>  529 :         current_statement__ = 1;
#>  530 :         w1_flat__ = context__.vals_r("w1");
#>  531 :         current_statement__ = 1;
#>  532 :         pos__ = 1;
#>  533 :         current_statement__ = 1;
#>  534 :         for (int sym1__ = 1; sym1__ <= nodes; ++sym1__) {
#>  535 :           current_statement__ = 1;
#>  536 :           for (int sym2__ = 1; sym2__ <= m; ++sym2__) {
#>  537 :             current_statement__ = 1;
#>  538 :             stan::model::assign(w1, w1_flat__[(pos__ - 1)],
#>  539 :               "assigning variable w1", stan::model::index_uni(sym2__),
#>  540 :               stan::model::index_uni(sym1__));
#>  541 :             current_statement__ = 1;
#>  542 :             pos__ = (pos__ + 1);
#>  543 :           }
#>  544 :         }
#>  545 :       }
#>  546 :       out__.write(w1);
#>  547 :       Eigen::Matrix<local_scalar_t__,-1,1> b1 =
#>  548 :         Eigen::Matrix<local_scalar_t__,-1,1>::Constant(nodes, DUMMY_VAR__);
#>  549 :       {
#>  550 :         std::vector<local_scalar_t__> b1_flat__;
#>  551 :         current_statement__ = 2;
#>  552 :         b1_flat__ = context__.vals_r("b1");
#>  553 :         current_statement__ = 2;
#>  554 :         pos__ = 1;
#>  555 :         current_statement__ = 2;
#>  556 :         for (int sym1__ = 1; sym1__ <= nodes; ++sym1__) {
#>  557 :           current_statement__ = 2;
#>  558 :           stan::model::assign(b1, b1_flat__[(pos__ - 1)],
#>  559 :             "assigning variable b1", stan::model::index_uni(sym1__));
#>  560 :           current_statement__ = 2;
#>  561 :           pos__ = (pos__ + 1);
#>  562 :         }
#>  563 :       }
#>  564 :       out__.write(b1);
#>  565 :       Eigen::Matrix<local_scalar_t__,-1,1> w_out =
#>  566 :         Eigen::Matrix<local_scalar_t__,-1,1>::Constant(nodes, DUMMY_VAR__);
#>  567 :       {
#>  568 :         std::vector<local_scalar_t__> w_out_flat__;
#>  569 :         current_statement__ = 3;
#>  570 :         w_out_flat__ = context__.vals_r("w_out");
#>  571 :         current_statement__ = 3;
#>  572 :         pos__ = 1;
#>  573 :         current_statement__ = 3;
#>  574 :         for (int sym1__ = 1; sym1__ <= nodes; ++sym1__) {
#>  575 :           current_statement__ = 3;
#>  576 :           stan::model::assign(w_out, w_out_flat__[(pos__ - 1)],
#>  577 :             "assigning variable w_out", stan::model::index_uni(sym1__));
#>  578 :           current_statement__ = 3;
#>  579 :           pos__ = (pos__ + 1);
#>  580 :         }
#>  581 :       }
#>  582 :       out__.write(w_out);
#>  583 :       local_scalar_t__ b_out = DUMMY_VAR__;
#>  584 :       current_statement__ = 4;
#>  585 :       b_out = context__.vals_r("b_out")[(1 - 1)];
#>  586 :       out__.write(b_out);
#>  587 :       local_scalar_t__ sigma = DUMMY_VAR__;
#>  588 :       current_statement__ = 5;
#>  589 :       sigma = context__.vals_r("sigma")[(1 - 1)];
#>  590 :       out__.write_free_lb(0, sigma);
#>  591 :     } catch (const std::exception& e) {
#>  592 :       stan::lang::rethrow_located(e, locations_array__[current_statement__]);
#>  593 :     }
#>  594 :   }
#>  595 :   inline void
#>  596 :   get_param_names(std::vector<std::string>& names__, const bool
#>  597 :                   emit_transformed_parameters__ = true, const bool
#>  598 :                   emit_generated_quantities__ = true) const {
#>  599 :     names__ = std::vector<std::string>{"w1", "b1", "w_out", "b_out", "sigma"};
#>  600 :     if (emit_transformed_parameters__) {
#>  601 :       std::vector<std::string> temp{"z1", "a1", "y_hat"};
#>  602 :       names__.reserve(names__.size() + temp.size());
#>  603 :       names__.insert(names__.end(), temp.begin(), temp.end());
#>  604 :     }
#>  605 :     if (emit_generated_quantities__) {}
#>  606 :   }
#>  607 :   inline void
#>  608 :   get_dims(std::vector<std::vector<size_t>>& dimss__, const bool
#>  609 :            emit_transformed_parameters__ = true, const bool
#>  610 :            emit_generated_quantities__ = true) const {
#>  611 :     dimss__ = std::vector<std::vector<size_t>>{std::vector<size_t>{static_cast<
#>  612 :                                                                     size_t>(m),
#>  613 :                                                  static_cast<size_t>(nodes)},
#>  614 :                 std::vector<size_t>{static_cast<size_t>(nodes)},
#>  615 :                 std::vector<size_t>{static_cast<size_t>(nodes)},
#>  616 :                 std::vector<size_t>{}, std::vector<size_t>{}};
#>  617 :     if (emit_transformed_parameters__) {
#>  618 :       std::vector<std::vector<size_t>>
#>  619 :         temp{std::vector<size_t>{static_cast<size_t>(n),
#>  620 :                static_cast<size_t>(nodes)},
#>  621 :              std::vector<size_t>{static_cast<size_t>(n),
#>  622 :                static_cast<size_t>(nodes)},
#>  623 :              std::vector<size_t>{static_cast<size_t>(n)}};
#>  624 :       dimss__.reserve(dimss__.size() + temp.size());
#>  625 :       dimss__.insert(dimss__.end(), temp.begin(), temp.end());
#>  626 :     }
#>  627 :     if (emit_generated_quantities__) {}
#>  628 :   }
#>  629 :   inline void
#>  630 :   constrained_param_names(std::vector<std::string>& param_names__, bool
#>  631 :                           emit_transformed_parameters__ = true, bool
#>  632 :                           emit_generated_quantities__ = true) const final {
#>  633 :     for (int sym1__ = 1; sym1__ <= nodes; ++sym1__) {
#>  634 :       for (int sym2__ = 1; sym2__ <= m; ++sym2__) {
#>  635 :         param_names__.emplace_back(std::string() + "w1" + '.' +
#>  636 :           std::to_string(sym2__) + '.' + std::to_string(sym1__));
#>  637 :       }
#>  638 :     }
#>  639 :     for (int sym1__ = 1; sym1__ <= nodes; ++sym1__) {
#>  640 :       param_names__.emplace_back(std::string() + "b1" + '.' +
#>  641 :         std::to_string(sym1__));
#>  642 :     }
#>  643 :     for (int sym1__ = 1; sym1__ <= nodes; ++sym1__) {
#>  644 :       param_names__.emplace_back(std::string() + "w_out" + '.' +
#>  645 :         std::to_string(sym1__));
#>  646 :     }
#>  647 :     param_names__.emplace_back(std::string() + "b_out");
#>  648 :     param_names__.emplace_back(std::string() + "sigma");
#>  649 :     if (emit_transformed_parameters__) {
#>  650 :       for (int sym1__ = 1; sym1__ <= nodes; ++sym1__) {
#>  651 :         for (int sym2__ = 1; sym2__ <= n; ++sym2__) {
#>  652 :           param_names__.emplace_back(std::string() + "z1" + '.' +
#>  653 :             std::to_string(sym2__) + '.' + std::to_string(sym1__));
#>  654 :         }
#>  655 :       }
#>  656 :       for (int sym1__ = 1; sym1__ <= nodes; ++sym1__) {
#>  657 :         for (int sym2__ = 1; sym2__ <= n; ++sym2__) {
#>  658 :           param_names__.emplace_back(std::string() + "a1" + '.' +
#>  659 :             std::to_string(sym2__) + '.' + std::to_string(sym1__));
#>  660 :         }
#>  661 :       }
#>  662 :       for (int sym1__ = 1; sym1__ <= n; ++sym1__) {
#>  663 :         param_names__.emplace_back(std::string() + "y_hat" + '.' +
#>  664 :           std::to_string(sym1__));
#>  665 :       }
#>  666 :     }
#>  667 :     if (emit_generated_quantities__) {}
#>  668 :   }
#>  669 :   inline void
#>  670 :   unconstrained_param_names(std::vector<std::string>& param_names__, bool
#>  671 :                             emit_transformed_parameters__ = true, bool
#>  672 :                             emit_generated_quantities__ = true) const final {
#>  673 :     for (int sym1__ = 1; sym1__ <= nodes; ++sym1__) {
#>  674 :       for (int sym2__ = 1; sym2__ <= m; ++sym2__) {
#>  675 :         param_names__.emplace_back(std::string() + "w1" + '.' +
#>  676 :           std::to_string(sym2__) + '.' + std::to_string(sym1__));
#>  677 :       }
#>  678 :     }
#>  679 :     for (int sym1__ = 1; sym1__ <= nodes; ++sym1__) {
#>  680 :       param_names__.emplace_back(std::string() + "b1" + '.' +
#>  681 :         std::to_string(sym1__));
#>  682 :     }
#>  683 :     for (int sym1__ = 1; sym1__ <= nodes; ++sym1__) {
#>  684 :       param_names__.emplace_back(std::string() + "w_out" + '.' +
#>  685 :         std::to_string(sym1__));
#>  686 :     }
#>  687 :     param_names__.emplace_back(std::string() + "b_out");
#>  688 :     param_names__.emplace_back(std::string() + "sigma");
#>  689 :     if (emit_transformed_parameters__) {
#>  690 :       for (int sym1__ = 1; sym1__ <= nodes; ++sym1__) {
#>  691 :         for (int sym2__ = 1; sym2__ <= n; ++sym2__) {
#>  692 :           param_names__.emplace_back(std::string() + "z1" + '.' +
#>  693 :             std::to_string(sym2__) + '.' + std::to_string(sym1__));
#>  694 :         }
#>  695 :       }
#>  696 :       for (int sym1__ = 1; sym1__ <= nodes; ++sym1__) {
#>  697 :         for (int sym2__ = 1; sym2__ <= n; ++sym2__) {
#>  698 :           param_names__.emplace_back(std::string() + "a1" + '.' +
#>  699 :             std::to_string(sym2__) + '.' + std::to_string(sym1__));
#>  700 :         }
#>  701 :       }
#>  702 :       for (int sym1__ = 1; sym1__ <= n; ++sym1__) {
#>  703 :         param_names__.emplace_back(std::string() + "y_hat" + '.' +
#>  704 :           std::to_string(sym1__));
#>  705 :       }
#>  706 :     }
#>  707 :     if (emit_generated_quantities__) {}
#>  708 :   }
#>  709 :   inline std::string get_constrained_sizedtypes() const {
#>  710 :     return std::string("[{\"name\":\"w1\",\"type\":{\"name\":\"matrix\",\"rows\":" + std::to_string(m) + ",\"cols\":" + std::to_string(nodes) + "},\"block\":\"parameters\"},{\"name\":\"b1\",\"type\":{\"name\":\"vector\",\"length\":" + std::to_string(nodes) + "},\"block\":\"parameters\"},{\"name\":\"w_out\",\"type\":{\"name\":\"vector\",\"length\":" + std::to_string(nodes) + "},\"block\":\"parameters\"},{\"name\":\"b_out\",\"type\":{\"name\":\"real\"},\"block\":\"parameters\"},{\"name\":\"sigma\",\"type\":{\"name\":\"real\"},\"block\":\"parameters\"},{\"name\":\"z1\",\"type\":{\"name\":\"matrix\",\"rows\":" + std::to_string(n) + ",\"cols\":" + std::to_string(nodes) + "},\"block\":\"transformed_parameters\"},{\"name\":\"a1\",\"type\":{\"name\":\"matrix\",\"rows\":" + std::to_string(n) + ",\"cols\":" + std::to_string(nodes) + "},\"block\":\"transformed_parameters\"},{\"name\":\"y_hat\",\"type\":{\"name\":\"vector\",\"length\":" + std::to_string(n) + "},\"block\":\"transformed_parameters\"}]");
#>  711 :   }
#>  712 :   inline std::string get_unconstrained_sizedtypes() const {
#>  713 :     return std::string("[{\"name\":\"w1\",\"type\":{\"name\":\"matrix\",\"rows\":" + std::to_string(m) + ",\"cols\":" + std::to_string(nodes) + "},\"block\":\"parameters\"},{\"name\":\"b1\",\"type\":{\"name\":\"vector\",\"length\":" + std::to_string(nodes) + "},\"block\":\"parameters\"},{\"name\":\"w_out\",\"type\":{\"name\":\"vector\",\"length\":" + std::to_string(nodes) + "},\"block\":\"parameters\"},{\"name\":\"b_out\",\"type\":{\"name\":\"real\"},\"block\":\"parameters\"},{\"name\":\"sigma\",\"type\":{\"name\":\"real\"},\"block\":\"parameters\"},{\"name\":\"z1\",\"type\":{\"name\":\"matrix\",\"rows\":" + std::to_string(n) + ",\"cols\":" + std::to_string(nodes) + "},\"block\":\"transformed_parameters\"},{\"name\":\"a1\",\"type\":{\"name\":\"matrix\",\"rows\":" + std::to_string(n) + ",\"cols\":" + std::to_string(nodes) + "},\"block\":\"transformed_parameters\"},{\"name\":\"y_hat\",\"type\":{\"name\":\"vector\",\"length\":" + std::to_string(n) + "},\"block\":\"transformed_parameters\"}]");
#>  714 :   }
#>  715 :   // Begin method overload boilerplate
#>  716 :   template <typename RNG> inline void
#>  717 :   write_array(RNG& base_rng, Eigen::Matrix<double,-1,1>& params_r,
#>  718 :               Eigen::Matrix<double,-1,1>& vars, const bool
#>  719 :               emit_transformed_parameters = true, const bool
#>  720 :               emit_generated_quantities = true, std::ostream*
#>  721 :               pstream = nullptr) const {
#>  722 :     const size_t num_params__ = (((((m * nodes) + nodes) + nodes) + 1) + 1);
#>  723 :     const size_t num_transformed = emit_transformed_parameters * ((((n *
#>  724 :       nodes) + (n * nodes)) + n));
#>  725 :     const size_t num_gen_quantities = emit_generated_quantities * (0);
#>  726 :     const size_t num_to_write = num_params__ + num_transformed +
#>  727 :       num_gen_quantities;
#>  728 :     std::vector<int> params_i;
#>  729 :     vars = Eigen::Matrix<double,-1,1>::Constant(num_to_write,
#>  730 :              std::numeric_limits<double>::quiet_NaN());
#>  731 :     write_array_impl(base_rng, params_r, params_i, vars,
#>  732 :       emit_transformed_parameters, emit_generated_quantities, pstream);
#>  733 :   }
#>  734 :   template <typename RNG> inline void
#>  735 :   write_array(RNG& base_rng, std::vector<double>& params_r, std::vector<int>&
#>  736 :               params_i, std::vector<double>& vars, bool
#>  737 :               emit_transformed_parameters = true, bool
#>  738 :               emit_generated_quantities = true, std::ostream*
#>  739 :               pstream = nullptr) const {
#>  740 :     const size_t num_params__ = (((((m * nodes) + nodes) + nodes) + 1) + 1);
#>  741 :     const size_t num_transformed = emit_transformed_parameters * ((((n *
#>  742 :       nodes) + (n * nodes)) + n));
#>  743 :     const size_t num_gen_quantities = emit_generated_quantities * (0);
#>  744 :     const size_t num_to_write = num_params__ + num_transformed +
#>  745 :       num_gen_quantities;
#>  746 :     vars = std::vector<double>(num_to_write,
#>  747 :              std::numeric_limits<double>::quiet_NaN());
#>  748 :     write_array_impl(base_rng, params_r, params_i, vars,
#>  749 :       emit_transformed_parameters, emit_generated_quantities, pstream);
#>  750 :   }
#>  751 :   template <bool propto__, bool jacobian__, typename T_> inline T_
#>  752 :   log_prob(Eigen::Matrix<T_,-1,1>& params_r, std::ostream* pstream = nullptr) const {
#>  753 :     Eigen::Matrix<int,-1,1> params_i;
#>  754 :     return log_prob_impl<propto__, jacobian__>(params_r, params_i, pstream);
#>  755 :   }
#>  756 :   template <bool propto__, bool jacobian__, typename T_> inline T_
#>  757 :   log_prob(std::vector<T_>& params_r, std::vector<int>& params_i,
#>  758 :            std::ostream* pstream = nullptr) const {
#>  759 :     return log_prob_impl<propto__, jacobian__>(params_r, params_i, pstream);
#>  760 :   }
#>  761 :   inline void
#>  762 :   transform_inits(const stan::io::var_context& context,
#>  763 :                   Eigen::Matrix<double,-1,1>& params_r, std::ostream*
#>  764 :                   pstream = nullptr) const final {
#>  765 :     std::vector<double> params_r_vec(params_r.size());
#>  766 :     std::vector<int> params_i;
#>  767 :     transform_inits(context, params_i, params_r_vec, pstream);
#>  768 :     params_r = Eigen::Map<Eigen::Matrix<double,-1,1>>(params_r_vec.data(),
#>  769 :                  params_r_vec.size());
#>  770 :   }
#>  771 :   inline void
#>  772 :   transform_inits(const stan::io::var_context& context, std::vector<int>&
#>  773 :                   params_i, std::vector<double>& vars, std::ostream*
#>  774 :                   pstream__ = nullptr) const {
#>  775 :     vars.resize(num_params_r__);
#>  776 :     transform_inits_impl(context, vars, pstream__);
#>  777 :   }
#>  778 :   inline void
#>  779 :   unconstrain_array(const std::vector<double>& params_constrained,
#>  780 :                     std::vector<double>& params_unconstrained, std::ostream*
#>  781 :                     pstream = nullptr) const {
#>  782 :     const std::vector<int> params_i;
#>  783 :     params_unconstrained = std::vector<double>(num_params_r__,
#>  784 :                              std::numeric_limits<double>::quiet_NaN());
#>  785 :     unconstrain_array_impl(params_constrained, params_i,
#>  786 :       params_unconstrained, pstream);
#>  787 :   }
#>  788 :   inline void
#>  789 :   unconstrain_array(const Eigen::Matrix<double,-1,1>& params_constrained,
#>  790 :                     Eigen::Matrix<double,-1,1>& params_unconstrained,
#>  791 :                     std::ostream* pstream = nullptr) const {
#>  792 :     const std::vector<int> params_i;
#>  793 :     params_unconstrained = Eigen::Matrix<double,-1,1>::Constant(num_params_r__,
#>  794 :                              std::numeric_limits<double>::quiet_NaN());
#>  795 :     unconstrain_array_impl(params_constrained, params_i,
#>  796 :       params_unconstrained, pstream);
#>  797 :   }
#>  798 : };
#>  799 : }
#>  800 : using stan_model = model1bee24d635d2__namespace::model1bee24d635d2_;
#>  801 : #ifndef USING_R
#>  802 : // Boilerplate
#>  803 : stan::model::model_base&
#>  804 : new_model(stan::io::var_context& data_context, unsigned int seed,
#>  805 :           std::ostream* msg_stream) {
#>  806 :   stan_model* m = new stan_model(data_context, seed, msg_stream);
#>  807 :   return *m;
#>  808 : }
#>  809 : stan::math::profile_map& get_stan_profile_data() {
#>  810 :   return model1bee24d635d2__namespace::profiles__;
#>  811 : }
#>  812 : #endif
#>  813 : #endif
#>  814 : 
#>  815 : RCPP_MODULE(stan_fit4model1bee24d635d2__mod) {
#>  816 :   class_<rstan::stan_fit<stan_model, boost::random::ecuyer1988> >(
#>  817 :       "stan_fit4model1bee24d635d2_")
#>  818 : 
#>  819 :       .constructor<SEXP, SEXP, SEXP>()
#>  820 : 
#>  821 :       .method(
#>  822 :           "call_sampler",
#>  823 :           &rstan::stan_fit<stan_model, boost::random::ecuyer1988>::call_sampler)
#>  824 :       .method(
#>  825 :           "param_names",
#>  826 :           &rstan::stan_fit<stan_model, boost::random::ecuyer1988>::param_names)
#>  827 :       .method("param_names_oi",
#>  828 :               &rstan::stan_fit<stan_model,
#>  829 :                                boost::random::ecuyer1988>::param_names_oi)
#>  830 :       .method("param_fnames_oi",
#>  831 :               &rstan::stan_fit<stan_model,
#>  832 :                                boost::random::ecuyer1988>::param_fnames_oi)
#>  833 :       .method(
#>  834 :           "param_dims",
#>  835 :           &rstan::stan_fit<stan_model, boost::random::ecuyer1988>::param_dims)
#>  836 :       .method("param_dims_oi",
#>  837 :               &rstan::stan_fit<stan_model,
#>  838 :                                boost::random::ecuyer1988>::param_dims_oi)
#>  839 :       .method("update_param_oi",
#>  840 :               &rstan::stan_fit<stan_model,
#>  841 :                                boost::random::ecuyer1988>::update_param_oi)
#>  842 :       .method("param_oi_tidx",
#>  843 :               &rstan::stan_fit<stan_model,
#>  844 :                                boost::random::ecuyer1988>::param_oi_tidx)
#>  845 :       .method("grad_log_prob",
#>  846 :               &rstan::stan_fit<stan_model,
#>  847 :                                boost::random::ecuyer1988>::grad_log_prob)
#>  848 :       .method("log_prob",
#>  849 :               &rstan::stan_fit<stan_model, boost::random::ecuyer1988>::log_prob)
#>  850 :       .method("unconstrain_pars",
#>  851 :               &rstan::stan_fit<stan_model,
#>  852 :                                boost::random::ecuyer1988>::unconstrain_pars)
#>  853 :       .method("constrain_pars",
#>  854 :               &rstan::stan_fit<stan_model,
#>  855 :                                boost::random::ecuyer1988>::constrain_pars)
#>  856 :       .method(
#>  857 :           "num_pars_unconstrained",
#>  858 :           &rstan::stan_fit<stan_model,
#>  859 :                            boost::random::ecuyer1988>::num_pars_unconstrained)
#>  860 :       .method(
#>  861 :           "unconstrained_param_names",
#>  862 :           &rstan::stan_fit<
#>  863 :               stan_model, boost::random::ecuyer1988>::unconstrained_param_names)
#>  864 :       .method(
#>  865 :           "constrained_param_names",
#>  866 :           &rstan::stan_fit<stan_model,
#>  867 :                            boost::random::ecuyer1988>::constrained_param_names)
#>  868 :       .method("standalone_gqs",
#>  869 :               &rstan::stan_fit<stan_model,
#>  870 :                                boost::random::ecuyer1988>::standalone_gqs);
#>  871 : }
#>  872 : 
#>  873 : 
#>  874 : // declarations
#>  875 : extern "C" {
#>  876 : SEXP file1bee263eaa9f( ) ;
#>  877 : }
#>  878 : 
#>  879 : // definition
#>  880 : SEXP file1bee263eaa9f() {
#>  881 :  return Rcpp::wrap("anon_model");
#>  882 : }
#> 
#> CHECKING DATA AND PREPROCESSING FOR MODEL 'anon_model' NOW.
#> 
#> COMPILING MODEL 'anon_model' NOW.
#> 
#> STARTING SAMPLER FOR MODEL 'anon_model' NOW.

# Access Stan model fit
model$fit
#> Inference for Stan model: anon_model.
#> 1 chains, each with iter=100; warmup=50; thin=1; 
#> post-warmup draws per chain=50, total post-warmup draws=50.
#> 
#>           mean se_mean   sd   2.5%   25%   50%   75% 97.5% n_eff Rhat
#> w1[1,1]  -0.12    0.15 0.95  -1.81 -0.67 -0.17  0.47  1.70    42 1.02
#> w1[1,2]  -0.12    0.14 1.19  -2.48 -0.92 -0.09  0.75  1.86    72 0.98
#> w1[2,1]  -0.10    0.14 1.01  -1.72 -0.83 -0.15  0.58  1.64    54 0.98
#> w1[2,2]  -0.18    0.14 1.19  -2.50 -0.79 -0.14  0.71  1.72    73 0.99
#> b1[1]    -0.20    0.12 0.95  -1.75 -0.96 -0.20  0.38  1.61    65 1.01
#> b1[2]    -0.11    0.18 1.04  -1.98 -0.86 -0.22  0.55  1.79    35 1.01
#> w_out[1]  0.15    0.15 1.08  -1.99 -0.53  0.01  0.62  2.13    54 0.98
#> w_out[2]  0.17    0.09 0.87  -1.69 -0.12  0.23  0.67  1.72    85 0.98
#> b_out     0.10    0.08 0.63  -1.07 -0.25  0.15  0.45  1.46    65 1.00
#> sigma     0.77    0.02 0.20   0.45  0.61  0.75  0.89  1.14    85 0.98
#> lp__     -6.93    0.60 2.29 -10.76 -8.35 -6.48 -5.08 -3.86    15 1.07
#> 
#> Samples were drawn using NUTS(diag_e) at Fri Dec 27 19:44:54 2024.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).