Print Method for "bnns" Objects

Displays a summary of a fitted Bayesian Neural Network (BNN) model, including the function call and the Stan fit details.

Usage

# S3 method for class 'bnns'
print(x, ...)

Arguments

x

An object of class "bnns", typically the result of a call to bnns.default.

...

Additional arguments (currently not used).

Value

The function is called for its side effects and does not return a value. It prints the following:

The function call used to generate the "bnns" object.

A summary of the Stan fit object stored in x$fit.

See also

Examples

# Example usage:
data <- data.frame(x1 = runif(10), x2 = runif(10), y = rnorm(10))
model <- bnns(y ~ -1 + x1 + x2,
  data = data, L = 1, nodes = 2, act_fn = 2,
  iter = 1e2, warmup = 5e1, chains = 1
)
#> 
#> SAMPLING FOR MODEL 'anon_model' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 1.3e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.13 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: WARNING: There aren't enough warmup iterations to fit the
#> Chain 1:          three stages of adaptation as currently configured.
#> Chain 1:          Reducing each adaptation stage to 15%/75%/10% of
#> Chain 1:          the given number of warmup iterations:
#> Chain 1:            init_buffer = 7
#> Chain 1:            adapt_window = 38
#> Chain 1:            term_buffer = 5
#> Chain 1: 
#> Chain 1: Iteration:  1 / 100 [  1%]  (Warmup)
#> Chain 1: Iteration: 10 / 100 [ 10%]  (Warmup)
#> Chain 1: Iteration: 20 / 100 [ 20%]  (Warmup)
#> Chain 1: Iteration: 30 / 100 [ 30%]  (Warmup)
#> Chain 1: Iteration: 40 / 100 [ 40%]  (Warmup)
#> Chain 1: Iteration: 50 / 100 [ 50%]  (Warmup)
#> Chain 1: Iteration: 51 / 100 [ 51%]  (Sampling)
#> Chain 1: Iteration: 60 / 100 [ 60%]  (Sampling)
#> Chain 1: Iteration: 70 / 100 [ 70%]  (Sampling)
#> Chain 1: Iteration: 80 / 100 [ 80%]  (Sampling)
#> Chain 1: Iteration: 90 / 100 [ 90%]  (Sampling)
#> Chain 1: Iteration: 100 / 100 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.006 seconds (Warm-up)
#> Chain 1:                0.005 seconds (Sampling)
#> Chain 1:                0.011 seconds (Total)
#> Chain 1: 
print(model)
#> Call:
#> bnns.default(formula = y ~ -1 + x1 + x2, data = data, L = 1, 
#>     nodes = 2, act_fn = 2, iter = 100, warmup = 50, chains = 1)
#> 
#> Stan fit:
#> Inference for Stan model: anon_model.
#> 1 chains, each with iter=100; warmup=50; thin=1; 
#> post-warmup draws per chain=50, total post-warmup draws=50.
#> 
#>            mean se_mean   sd   2.5%    25%    50%   75% 97.5% n_eff Rhat
#> w1[1,1]    0.09    0.19 1.21  -1.55  -1.00   0.12  0.87  2.56    40 1.03
#> w1[1,2]   -0.06    0.20 1.39  -2.66  -0.97   0.00  0.96  2.42    50 1.02
#> w1[2,1]   -0.03    0.13 0.85  -2.05  -0.55   0.02  0.46  1.27    41 0.98
#> w1[2,2]   -0.01    0.15 0.93  -1.65  -0.78   0.12  0.66  1.47    40 1.01
#> b1[1]     -0.21    0.16 0.84  -1.94  -0.64  -0.26  0.32  1.38    27 1.14
#> b1[2]     -0.11    0.12 0.82  -1.35  -0.71  -0.17  0.41  1.50    51 0.98
#> w_out[1]   0.08    0.17 1.16  -1.96  -0.72  -0.02  0.97  2.07    49 0.98
#> w_out[2]   0.05    0.19 0.98  -1.59  -0.44  -0.05  0.77  1.78    28 0.99
#> b_out      0.30    0.09 0.74  -1.01  -0.24   0.19  0.90  1.51    63 0.99
#> sigma      1.07    0.03 0.23   0.69   0.95   1.05  1.18  1.56    49 0.98
#> lp__     -11.00    0.35 2.10 -15.17 -12.33 -11.11 -9.55 -7.33    36 0.99
#> 
#> Samples were drawn using NUTS(diag_e) at Tue Jan  7 10:57:15 2025.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).