Summary of a Bayesian Neural Network (BNN) Model

Provides a comprehensive summary of a fitted Bayesian Neural Network (BNN) model, including details about the model call, data, network architecture, posterior distributions, and model fitting information.

Usage

# S3 method for class 'bnns'
summary(object, ...)

Arguments

object

An object of class bnns, representing a fitted Bayesian Neural Network model.

...

Additional arguments (currently unused).

Value

A list (returned invisibly) containing the following elements:

"Number of observations": The number of observations in the training data.

"Number of features": The number of features in the training data.

"Number of hidden layers": The number of hidden layers in the neural network.

"Nodes per layer": A comma-separated string representing the number of nodes in each hidden layer.

"Activation functions": A comma-separated string representing the activation functions used in each hidden layer.

"Output activation function": The activation function used in the output layer.

"Stanfit Summary": A summary of the Stan model, including key parameter posterior distributions.

"Iterations": The total number of iterations used for sampling in the Bayesian model.

"Warmup": The number of iterations used as warmup in the Bayesian model.

"Thinning": The thinning interval used in the Bayesian model.

"Chains": The number of Markov chains used in the Bayesian model.

"Performance": Predictive performance metrics, which vary based on the output activation function.

The function also prints the summary to the console.

Details

The function prints the following information:

Call: The original function call used to fit the model.

Data Summary: Number of observations and features in the training data.

Network Architecture: Structure of the BNN including the number of hidden layers, nodes per layer, and activation functions.

Posterior Summary: Summarized posterior distributions of key parameters (e.g., weights, biases, and noise parameter).

Model Fit Information: Bayesian sampling details, including the number of iterations, warmup period, thinning, and chains.

Notes: Remarks and warnings, such as checks for convergence diagnostics.

See also

Examples

# \donttest{
# Fit a Bayesian Neural Network
data <- data.frame(x1 = runif(10), x2 = runif(10), y = rnorm(10))
model <- bnns(y ~ -1 + x1 + x2,
  data = data, L = 1, nodes = 2, act_fn = 2,
  iter = 1e2, warmup = 5e1, chains = 1
)
#> 
#> SAMPLING FOR MODEL 'anon_model' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 1.2e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.12 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: WARNING: There aren't enough warmup iterations to fit the
#> Chain 1:          three stages of adaptation as currently configured.
#> Chain 1:          Reducing each adaptation stage to 15%/75%/10% of
#> Chain 1:          the given number of warmup iterations:
#> Chain 1:            init_buffer = 7
#> Chain 1:            adapt_window = 38
#> Chain 1:            term_buffer = 5
#> Chain 1: 
#> Chain 1: Iteration:  1 / 100 [  1%]  (Warmup)
#> Chain 1: Iteration: 10 / 100 [ 10%]  (Warmup)
#> Chain 1: Iteration: 20 / 100 [ 20%]  (Warmup)
#> Chain 1: Iteration: 30 / 100 [ 30%]  (Warmup)
#> Chain 1: Iteration: 40 / 100 [ 40%]  (Warmup)
#> Chain 1: Iteration: 50 / 100 [ 50%]  (Warmup)
#> Chain 1: Iteration: 51 / 100 [ 51%]  (Sampling)
#> Chain 1: Iteration: 60 / 100 [ 60%]  (Sampling)
#> Chain 1: Iteration: 70 / 100 [ 70%]  (Sampling)
#> Chain 1: Iteration: 80 / 100 [ 80%]  (Sampling)
#> Chain 1: Iteration: 90 / 100 [ 90%]  (Sampling)
#> Chain 1: Iteration: 100 / 100 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.005 seconds (Warm-up)
#> Chain 1:                0.004 seconds (Sampling)
#> Chain 1:                0.009 seconds (Total)
#> Chain 1: 

# Get a summary of the model
summary(model)
#> Call:
#> bnns.default(formula = y ~ -1 + x1 + x2, data = data, L = 1, 
#>     nodes = 2, act_fn = 2, iter = 100, warmup = 50, chains = 1)
#> 
#> Data Summary:
#> Number of observations: 10 
#> Number of features: 2 
#> 
#> Network Architecture:
#> Number of hidden layers: 1 
#> Nodes per layer: 2 
#> Activation functions: 2 
#> Output activation function: 1 
#> 
#> Posterior Summary (Key Parameters):
#>                 mean    se_mean        sd       2.5%        25%         50%
#> w_out[1] -0.05432372 0.09247250 0.8053585 -1.8116323 -0.5900625 -0.01134891
#> w_out[2] -0.20667669 0.11063755 0.8241266 -1.6868328 -0.7247308 -0.15475959
#> b_out     0.22370414 0.08898919 0.6674141 -0.7754468 -0.1774560  0.16644801
#> sigma     1.17390514 0.02727134 0.2513531  0.8313935  1.0085272  1.09219326
#>                75%    97.5%    n_eff      Rhat
#> w_out[1] 0.5618538 1.316687 75.84960 0.9939791
#> w_out[2] 0.3165364 1.430570 55.48592 1.0217312
#> b_out    0.6322990 1.604595 56.24919 1.0572413
#> sigma    1.3415874 1.736803 84.94850 0.9825971
#> 
#> Model Fit Information:
#> Iterations: 100 
#> Warmup: 50 
#> Thinning: 1 
#> Chains: 1 
#> 
#> Predictive Performance:
#> RMSE (training): 1.1672 
#> MAE (training): 1.016994 
#> 
#> Notes:
#> Check convergence diagnostics for parameters with high R-hat values.
# }